% Generated by roxygen2 (4.0.2): do not edit by hand
\name{adj.duplicated}
\alias{adj.duplicated}
\title{base_duplicated but for adjacent only}
\usage{
adj.duplicated(vec)
}
\arguments{
\item{vec}{vector of values to test}
}
\value{
logical vector with same length as vec
}
\description{
Same as base duplicated function, but only counts duplicates that are adjacent.
e.g. duplicated(c(1,2,2,3,2)) returns c(F,F,T,F,T)
 duplicated.adj(c(1,2,2,3,2)) returns c(F,F,T,F,F)
}

