% Generated by roxygen2 (4.0.2): do not edit by hand
\name{alpha}
\alias{alpha}
\title{alpha}
\usage{
alpha(col, alpha)
}
\arguments{
\item{col}{color string}

\item{alpha}{alpha value between 0 and 1}
}
\description{
takes provided colors and gives them the specified alpha (ie transparency) value
}

