% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{anchorlift}
\alias{anchorlift}
\title{anchorlift}
\usage{
anchorlift(query, subject, window = 1e+09, by = NULL,
  seqname = "Anchor", include.values = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{query}{GRanges that will be lifted around the subject}

\item{subject}{GRanges around which the queries will be lifted}

\item{window}{integer specifying how far around each subject to gather query intervals to lift (default = 1e9)}

\item{by}{character vector specifying additional columms (e.g. sample id) around which to restrict overlaps (via gr.findoverlaps()). Refer to `gr.findoverlaps()` documentation. (default = NULL)}

\item{seqname}{String specifying the name of the output sequence around which to anchor (default = "Anchor")}

\item{include.values}{Boolean Flag whether to include values from query and subject (default = TRUE)}
}
\value{
anchorlifted GRanges
}
\description{
"lifts" all queries with respect to subject in coordinates that are within "pad"
i.e. puts the queries into subject-centric coordinates, which is a new genome with label "Anchor" (default)

Respects strand of subject (i.e. if subject strand gr is "-" then will lift all queries to the left of it
into positive subject-centric coordinates). Keeps
track of subject and query id for later deconvolution if need be.
}
\author{
Marcin Imielinski
}
