% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bam.cov.thresh}
\alias{bam.cov.thresh}
\title{Determines "covered" vs "noncovered" regions according to user specified quality threshold
Outputs GRanges object of covered regions}
\usage{
bam.cov.thresh(bam.file, thresh = 20, verbose = T, chunksize = 1e+05)
}
\arguments{
\item{bam.file}{character scalar input bam file}

\item{window}{specify window size}
}
\description{
Counts coverage in "window" bp of genome from BAM file "bam.file" using foghorn library
Writes bed file (.bed) and Granges file (.rds) to out.file path
}

