% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bam.cov.tile.st}
\alias{bam.cov.tile.st}
\title{Quick way to get tiled coverage via piping to samtools (~10 CPU-hours for 100bp tiles, 5e8 read pairs)}
\usage{
bam.cov.tile.st(bam.file, window = 100, chunksize = 1e+05, min.mapq = 30,
  verbose = T, max.tlen = 10000, st.flag = "-f 0x02 -F 0x10",
  fragments = TRUE, region = NULL, do.gc = FALSE, midpoint = T)
}
\arguments{
\item{bam.file}{character scalar input bam file}

\item{window}{integer scalar window size (in bp)}

\item{chunksize}{integer scalar, size of window}

\item{min.mapq}{integer scalar, minimim map quality reads to consider for counts}
}
\value{
GRanges of "window" bp tiles across seqlengths of bam.file with meta data field $counts specifying fragment counts centered
in the given bin.
}
\description{
Gets coverage for window size "window", pulling "chunksize" records at a time and incrementing bin
corresponding to midpoint or overlaps of corresponding (proper pair) fragment (uses TLEN and POS for positive strand reads that are part of a proper pair)
}

