% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{bamflag}
\alias{bamflag}
\alias{bamtag}
\title{bamtag}
\usage{
bamtag(reads, secondary = F, gr.string = F)
}
\arguments{
\item{reads}{GenomicRanges holding the reads}
}
\description{
outputs a tag that cats qname, first vs first second mate +/- secondary alignment +/- gr.string
to give an identifier for determine duplicates in a read pile
}

