% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{bamflagr}
\alias{bamflag}
\alias{bamflagr}
\title{bamflag}
\usage{
bamflag(reads)
}
\arguments{
\item{reads}{GenomicRanges holding the reads}
}
\description{
shortcut .. assumes reads are GappedAlignments with flag variable or actual integers representing bam flag
}

