% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{change_seqlevels_style}
\alias{change_seqlevels_style}
\title{Change seqlevels}
\usage{
change_seqlevels_style(gr, style = "NCBI")
}
\description{
Change seqlevels style
}
\details{
Wrapper to GenomeInfoDb functions to
change seqlevels style. "NCBI" = no chr,
"UCSC" = chr prefixed.
}
\author{
Kevin Hadi
}
