% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{chunk}
\alias{chunk}
\title{chunk}
\usage{
chunk(from, to = NULL, by = 1, length.out = NULL)
}
\arguments{
\item{from}{dummy}

\item{to}{dummy}

\item{by}{Default 1}

\item{length.out}{Default NULL}
}
\value{
2-column matrix of indices corresponding to "chunks"
}
\description{
takes same input as seq (from, to, by, length.out) and outputs a 2 column matrix of indices 
corresponding to "chunks"
}

