% Generated by roxygen2 (4.0.2): do not edit by hand
\name{count.clips}
\alias{count.clips}
\title{count.clips}
\usage{
count.clips(reads, hard = FALSE)
}
\arguments{
\item{reads}{GenomicRanges holding the reads}

\item{hard}{[Default TRUE] option counts hard clips}
}
\description{
takes gr or gappedalignment object and uses cigar field (or takes character vector of cigar strings)
and returns data frame with fields (for character input)
$right.clips number of "right" soft clips (eg cigar 89M12S)
#left.clips number of "left" soft clips (eg cigar 12S89M)
or appends these fields to the reads object
}

