% Generated by roxygen2 (4.0.2): do not edit by hand
\name{countCigar}
\alias{countCigar}
\title{Count bases in cigar string}
\usage{
countCigar(cigar)
}
\arguments{
\item{cigar}{character vector of cigar strings}
}
\value{
a 4-column, length(cigar)-row matrix with the total counts for each type
}
\description{
Counts the total number of bases, per cigar, that fall into D, I, M, S categories.
}

