% Generated by roxygen2 (4.0.2): do not edit by hand
\name{discordant.pairs}
\alias{discordant.pairs}
\title{Label discordant read pairs}
\usage{
discordant.pairs(pairs, inter.only = F, dmin = 50, dmax = 500)
}
\description{
Labels read pairs discordant based on whether they have (1) ++ or -- strand orientation (2) "-" strand read start
is not greater than dmin or less than dmin ahead of  "+" strand read on same chromosome
}
\note{
need to merge with gr.isdisc
}

