% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dt2gr}
\alias{dt2gr}
\title{Convert data.table to GRanges}
\usage{
dt2gr(dt)
}
\arguments{
\item{dt}{data.table to convert to GRanges}
}
\value{
GRanges object of length = nrow(dt)
}
\description{
Takes as input a data.table which must have the fields: start, end, strand, seqnames.
All of the remaining fields are added as meta data to the GRanges
}
\examples{
gr <- dt2gr(data.table(start=c(1,2), seqnames=c("X", "1"), end=c(10,20), strand = c('+', '-')))
}
\author{
Jeremiah Wal
}

