% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\docType{data}
\name{example_dnase}
\alias{example_dnase}
\title{DNAaseI hypersensitivity sites for hg19A}
\format{\code{GRanges}}
\description{
DNAaseI hypersensitivity sites from UCSC Table Browser hg19,
subsampled to 10,000 sites
}
\keyword{data}

