% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gc_content}
\alias{gc_content}
\title{gc_content}
\usage{
gc_content(segs, build = "hg19")
}
\description{
Uses BSGenome package to compute gc content for a collection of segments in seg data frame ($chr, $start, $end or $chr, $pos1, $pos2 or $chr, $begin, $end)
Returns vector of gc content of length nrow(segs).
}

