% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{get.mate.gr}
\alias{get.mate.gr}
\title{Create GRanges of read mates from reads}
\usage{
get.mate.gr(reads)
}
\value{
\code{GRanges} corresponding to mates of reads
}
\description{
Create GRanges of read mates from reads
}

