% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.pairs.grl}
\alias{get.pairs.grl}
\title{takes reads object and returns grl with each read and its mate (if exists)}
\usage{
get.pairs.grl(reads, as.grl = TRUE, verbose = F)
}
\description{
takes reads object and returns grl with each read and its mate (if exists)
}

