% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{get_seq}
\alias{get_seq}
\title{get_seq
Retrieve genomic sequenes

Wrapper around getSeq which does the "chr" and seqnames conversion if necessary
also handles GRangesList queries}
\usage{
get_seq(hg, gr, unlist = TRUE, mc.cores = 1, mc.chunks = mc.cores,
  as.data.table = FALSE, verbose = FALSE)
}
\arguments{
\item{hg}{A BSgenome or and ffTrack object with levels = c('A','T','G','C','N')}

\item{gr}{GRanges object to define the ranges}

\item{unlist}{logical whether to unlist the final output into a single DNAStringSet. Default TRUE}

\item{mc.cores}{Optional multicore call. Default 1}

\item{mc.chunks}{Optional define how to chunk the multicore call. Default mc.cores}

\item{verbose}{Increase verbosity}
}
\value{
DNAStringSet of sequences
}
\description{
get_seq
Retrieve genomic sequenes

Wrapper around getSeq which does the "chr" and seqnames conversion if necessary
also handles GRangesList queries
}

