% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave_pdf_png.R
\name{ggsave_pdf_png}
\alias{ggsave_pdf_png}
\title{Save a pdf and png image in the same directory with the same filename
Useful for creating identical versions of a publication figure}
\usage{
ggsave_pdf_png(p, path, width, height)
}
\arguments{
\item{p}{ggplot object, ggplot object to be saved}

\item{path}{character, output pdf filepath}

\item{width}{numeric, output figure width (inches)}

\item{height}{numeric, output figure height (inches)}
}
\value{
none
}
\description{
Save a pdf and png image in the same directory with the same filename
Useful for creating identical versions of a publication figure
}
