% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gr.addunmap}
\alias{gr.addunmap}
\title{In addition to adding the Unmapped seqlevel
to the GRanges, it can optionally modify the
seqnames and field \code{type} for any GRanges
level that starts at 1 (current marker for unmapped)}
\usage{
gr.addunmap(gr, add.type = F, readlen = 101)
}
\arguments{
\item{gr}{GRanges to modify}

\item{add.type}{add Type field and change seqlevels of Unmapped reads to Unmapped (see description). Default T}
}
\value{
GRanges with the modified seqlevels
}
\description{
In addition to adding the Unmapped seqlevel
to the GRanges, it can optionally modify the
seqnames and field \code{type} for any GRanges
level that starts at 1 (current marker for unmapped)
}
\examples{
##
}
\author{
Jeremiah Wala \email{jwala@broadinstitute.org}
}

