% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.all}
\alias{gr.all}
\title{gr.all}
\usage{
gr.all(unmap = FALSE, M = FALSE, Y = TRUE)
}
\arguments{
\item{unmap}{[default F] Optinally add a "unmapped" chr}

\item{M}{[default F] Include mitochondrial chr}

\item{Y}{[default T] Include Y chr}
}
\value{
\code{GRanges} object with one element per chromosome
}
\description{
Return a GRanges that holds interavals for all of HG19
}

