% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.bind}
\alias{gr.bind}
\title{Concatenate \code{GRanges}, robust to different \code{mcols}}
\usage{
gr.bind(x, ...)
}
\arguments{
\item{x}{GRanges input GRanges}

\item{...}{additional input GRanges}
}
\value{
Concatenated \code{GRanges}
gr.bind(example_genes, example_dnase)
}
\description{
Concatenates \code{GRanges} objects, taking the union of their features if they have non-overlapping features
}
\note{
Does not fill in the \code{Seqinfo} for the output \code{GRanges}
}
