% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.breaks}
\alias{gr.breaks}
\title{gr.breaks}
\usage{
gr.breaks(bps = NULL, query = NULL)
}
\arguments{
\item{bps}{\code{GRanges} of width 1, locations of the bp; if any element width
larger than 1, both boundary will be considered individual breakpoints}

\item{query}{a disjoint \code{GRanges} object to be broken}
}
\value{
\code{GRanges} disjoint object at least the same length as query,
with a metadata column \code{qid} indicating input index where new segment is from
}
\description{
Break GRanges at given breakpoints into disjoint gr
}
\author{
Xiaotong Yao
}
