% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.disjoin}
\alias{gr.disjoin}
\title{GenomicRanges disjoin with some additional functionality}
\usage{
gr.disjoin(x, ..., ignore.strand = TRUE)
}
\arguments{
\item{x}{GRanges to disjoin}

\item{...}{arguments to disjoin (e.g. with.revmap=FALSE). Please see documentation for GenomicRanges::disjoin()}

\item{ignore.strand}{logical scalar (default = TRUE)}
}
\value{
GRanges of non-overlapping ranges with metadata
}
\description{
Identical to GRanges disjoin, except outputs inherit metadata from first overlapping parent instance on input
}
