% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gr.dist}
\alias{gr.dist}
\title{gr.dist}
\usage{
gr.dist(gr1, gr2 = NULL, ignore.strand = T, ...)
}
\description{
computes pairwise distance between elements of two gr objects of length n and m
returning n by m matrix of distances between item i of gr1 and item j of gr2
}
\details{
distances are computed as follows:
NA for ranges on different seqnames
0 for overlapping ranges)
min(abs(end1-end2), abs(end1-start2), abs(start1-end2), abs(start1-end1),) for all others

if only gr1 is provided, then will return n x n matrix of gr's vs themselves

if max.dist = T then will replace min with max above
}

