% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.duplicated}
\alias{gr.duplicated}
\title{Allows to restrict duplicates using "by" columns and allows in exact matching}
\usage{
gr.duplicated(query, by = NULL, type = "any")
}
\arguments{
\item{query}{GRanges to query}

\item{by}{string Column 'by' used to restrict duplicates. See the 'by' argument for function gr.match()}

\item{type}{string 'type' used. See the 'type' argument for function gr.match()}
}
\value{
boolean vector of match status
}
\description{
Allows to restrict duplicates using "by" columns and allows in exact matching
}
