% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.fix}
\alias{gr.fix}
\title{"Fixes" seqlengths / seqlevels}
\usage{
gr.fix(gr, genome = NULL, gname = NULL, drop = FALSE)
}
\arguments{
\item{gr}{\code{GRanges} object to fix}

\item{genome}{Genome to fix to: \code{Seqinfo}, \code{BSgenome}, \code{GRanges} (w/seqlengths), \code{GRangesList} (w/seqlengths)}

\item{gname}{Name of the genome (optional, just appends to \code{Seqinfo} of the output) [NULL]}

\item{drop}{Remove ranges that are not present in the supplied genome [FALSE]}
}
\value{
\code{GRanges} pile with the fixed \code{Seqinfo}
}
\description{
If "genome" not specified will replace NA seq lengths in GR to reflect largest coordinate per seqlevel
and removes all NA seqlevels after this fix.
}
\details{
if "genome" defined (ie as Seqinfo object, or a BSgenome, GRanges, GRnagesList object with populated seqlengths) then will replace
seqlengths in gr with those for that genome (and if drop = T, drop all ranges without
seqlevels in that genome)
}

