% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.fixseq}
\alias{gr.fixseq}
\title{Get union of seqlengths from multiple GRanges-like objects}
\usage{
gr.fixseq(...)
}
\value{
integer vector
}
\description{
Creates union of seqlengths from arbitrary number of GenomicRanges objects
}
\author{
Kevin Hadi
}
