% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gr.flip}
\alias{gr.flip}
\title{gr.flip}
\usage{
gr.flip(gr, which = TRUE)
}
\arguments{
\item{gr}{\code{GRanges} to flip}

\item{which}{Default TRUE}
}
\description{
flips strand on grs
optional arg which will determine which ones to flip
}

