% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.flipstrand}
\alias{gr.flipstrand}
\title{Flip strand on \code{GRanges}}
\usage{
gr.flipstrand(gr)
}
\arguments{
\item{gr}{\code{GRanges} pile with strands to be flipped}
}
\value{
\code{GRanges} with flipped strands (+ to -, * to *, - to *)
}
\description{
Flip strand on \code{GRanges}
}
\examples{
gr.flipstrand(GRanges(1, IRanges(c(10,10,10),20), strand=c("+","*","-")))
}

