% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\docType{data}
\name{gr.genes}
\alias{gr.genes}
\title{RefSeq genes from UCSC Table Browser hg19, subsampled to 10,000 genes}
\format{\code{GRanges}}
\description{
RefSeq genes from UCSC Table Browser hg19, subsampled to 10,000 genes
}
\keyword{data}

