% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.genome}
\alias{gr.genome}
\title{create GRanges of full genome coordinates}
\usage{
gr.genome(si, onlystandard = TRUE, genome = NULL)
}
\value{
GRanges
}
\description{
Grabs *.chrom.sizes file from
environmental variable "DEFAULT_GENOME" or
"DEFAULT_BSGENOME"

May need to set either of these via
Sys.setenv(DEFAULT_BSGENOME = "path_to_ref.chrom.sizes")
Sys.setenv(DEFAULT_GENOME = "path_to_ref.chrom.sizes")
}
\author{
Kevin Hadi
}
