% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\docType{methods}
\name{\%^^\%}
\alias{\%^\%}
\alias{\%^\%,GRanges-method}
\alias{\%^^\%}
\alias{\%^^\%,GRanges-method}
\title{gr.in shortcut (respects strand)}
\usage{
x \%^^\% ...

x \%^\% ...
}
\arguments{
\item{x}{See \link{gr.in}}

\item{...}{See \link{gr.in}}

\item{x}{\code{GRanges} object}

\item{...}{additional arguments to gr.in}
}
\value{
logical vector of length gr1 which is TRUE at entry i only if gr1[i] intersects at least one interval in gr2

logical vector of length gr1 which is TRUE at entry i only if gr1[i] intersects at least one interval in gr2 (strand agnostic)
}
\description{
Shortcut for gr.in

gr1 %^^% gr2

Shortcut for gr.in (standard arguments)

gr1 %^% gr2
}
\author{
Marcin Imielinski
}

