% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\docType{methods}
\name{\%&&\%}
\alias{\%&&\%}
\alias{\%&&\%,GRanges-method}
\alias{\%&&\%,GRangesList-method}
\alias{\%&&\%,CompressedGRangesList-method}
\title{Subset x on y ranges, strand-specific}
\usage{
x \%&&\% y

\S4method{\%&&\%}{GRangesList}(x, y)

\S4method{\%&&\%}{CompressedGRangesList}(x, y)
}
\value{
subset of gr1 that overlaps gr2

subset of gr1 that overlaps gr2

subset of gr1 that overlaps gr2
}
\description{
shortcut for x[gr.in(x,y)]

gr1 %&&% gr2 returns the subsets of gr1 that overlaps gr2

shortcut for x[gr.in(x,y)]

gr1 %&&% gr2 returns the subsets of gr1 that overlaps gr2

shortcut for x[gr.in(x,y)]

gr1 %&&% gr2 returns the subsets of gr1 that overlaps gr2
}
\author{
Marcin Imielinski

Marcin Imielinski

Marcin Imielinski
}
