% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{\%WW\%}
\alias{\%WW\%}
\alias{\%WW\%,GRanges-method}
\title{subset x on y ranges wise obeying strand}
\usage{
x \%WW\% ...
}
\arguments{
\item{x}{See \link{gr.in}}

\item{...}{See \link{gr.in}}
}
\value{
subset of gr1 that overlaps gr2
}
\description{
shortcut for x[gr.in(x,y, ignore.strand = FALSE)]

gr1 %WW% gr2 returns the subsets of gr that overlaps gr2 not ignoring strand
}
\author{
Marcin Imielinski
}

