% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.in}
\alias{\%^\%}
\alias{gr.in}
\title{Versatile implementation of \code{GenomicRanges::over}}
\usage{
gr.in(query, subject, ...)

x \%^\% ...
}
\arguments{
\item{query}{\code{GRanges}}

\item{subject}{\code{GRanges}}

\item{...}{Argument to be sent to \code{\link{gr.findoverlaps}} (e.g. \code{by})}

\item{x}{GRanges object}

\item{...}{additional arguments to gr.in}
}
\value{
logical vector of length gr1 which is TRUE at entry i only if gr1[i] intersects at least one interval in gr2 (strand agnostic)
}
\description{
returns T / F vector if query range i is found in any range in subject

Shortcut for gr.in (standard arguments)

gr1 %^% gr2
}
\author{
Marcin Imielinski
}

