% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{\%&\%}
\alias{\%&&\%}
\alias{\%&\%}
\title{gr.intersect shortcut}
\usage{
x \%&\% ...

x \%&&\% ...
}
\value{
granges representing intersection of input intervals

granges representing intersection of input intervals
}
\description{
Shortcut for gr.intersect

gr1 %&% gr2

Shortcut for gr.intersect

gr1 %&&% gr2
}
\author{
Marcin Imielinski

Marcin Imielinski
}

