% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.isclip}
\alias{gr.isclip}
\title{Check if reads are clipped}
\usage{
gr.isclip(gr, clip.cutoff = 10)
}
\arguments{
\item{gr}{Granges OR data.table that has \code{cigar} field and \code{qname} field}

\item{clip.cutoff}{Minimum number of bases that are clipped to call the reads as clipped}
}
\value{
logical of length of input, denoting whether that read is part of a clipped read pair.
}
\description{
Returns a logical vector of length of the input GRanges that
that classifies a read as clipped or not. The user can specify
a cutoff value for how many bases need to be clipped.
}

