% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gr.isdisc}
\alias{gr.isdisc}
\title{Checks if reads are discordant}
\usage{
gr.isdisc(gr, isize = 1000, unmap.only = FALSE)
}
\arguments{
\item{gr}{Granges OR data.table that has \code{isize} field and \code{qname} field}

\item{isize}{Minimum insert size required to call discordant. Default 1000}

\item{unmap.only}{Find only pairs with an unmapped read}
}
\value{
logical vector of length of input, denoting each read as discordant or not
}
\description{
Returns a logical vector denoting if a read is discordant.
There is only a minimum absolute isize, and any read below this is
not considered discordant. This will return logicals based on read pairs
}

