% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gr.match}
\alias{gr.match}
\title{gr.match}
\usage{
gr.match(query, subject, max.slice = Inf, verbose = FALSE, mc.cores = 1,
  ...)
}
\description{
Faster implementation of GRanges match (uses gr.findoverlaps)
returns indices of query in subject or NA if none found
... = additional args for findOverlaps (IRanges version)
}

