% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.merge}
\alias{gr.merge}
\title{merge GRanges using coordinates as primary key}
\usage{
gr.merge(query, subject, by = NULL, all = FALSE, all.query = all,
  all.subject = all, ignore.strand = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{query}{Query \code{GRanges} pile}

\item{subject}{Subject \code{GRanges} pile}

\item{by}{vector Additional metadata fields to join on}

\item{all}{boolean Flag whether to include left and right joins}

\item{all.query}{whether to do a left join (default==all)}

\item{all.subject}{whether to do a right join (default==all)}
}
\description{
Uses gr.findoverlaps to enable internal and external joins of GRanges using
syntax similar to "merge" where merging is done using coordinates +/- "by" fields

Uses gr.findoverlaps / findOverlaps for heavy lifting, but returns outputs with
metadata populated as well as query and subject ids.  For external joins,
overlaps x with gaps(y) and gaps(x) with y.
}
