% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.mid}
\alias{gr.mid}
\title{Get the midpoints of \code{GRanges} ranges}
\usage{
gr.mid(x)
}
\arguments{
\item{x}{\code{GRanges} object to operate on}
}
\value{
\code{GRanges} of the midpoint, calculated from \code{floor(width(x)/2)}
}
\description{
Get the midpoints of \code{GRanges} ranges
}
\examples{
gr.mid(GRanges(1, IRanges(1000,2000), seqinfo=Seqinfo("1", 2000)))
}

