% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gr.nochr}
\alias{gr.nochr}
\title{Remove chr prefix from GRanges seqlevels}
\usage{
gr.nochr(gr)
}
\arguments{
\item{gr}{GRanges with chr seqlevel prefixes}
}
\value{
GRanges without chr seqlevel prefixes
}
\description{
Remove chr prefix from GRanges seqlevels
}
\examples{
##
}
\author{
Jeremiah Wala \email{jwala@broadinstitute.org}
}

