% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.pairflip}
\alias{gr.pairflip}
\title{gr.pairflip}
\usage{
gr.pairflip(gr)
}
\arguments{
\item{gr}{\code{GRanges}}
}
\description{
"pairs" gr returning a grl with each item consisting
of the original gr and its strand flip
}

