% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gr.peaks}
\alias{gr.peaks}
\title{gr.peaks}
\usage{
gr.peaks(gr, field = "score", minima = F, peel = 0, id.field = NULL,
  bootstrap = TRUE, na.rm = TRUE, pbootstrap = 0.95, nbootstrap = 10000,
  FUN = NULL, AGG.FUN = sum, peel.gr = NULL, score.only = FALSE,
  verbose = peel > 0)
}
\description{
Finds "peaks" in an input GRanges with value field y.
first piles up ranges according to field score (default = 1 for each range)
then finds peaks.  If peel > 0, then recursively peels segments
contributing to top peak, and recomputes nextpeak "peel" times
if peel>0, bootstrap controls whether to bootstrap peak interval nbootstrap times
if id.field is not NULL will peel off with respect to unique (sample) id of segment and not purely according to width
if FUN preovided then will complex aggregating function of piled up values in dijoint intervals prior to computing "coverage"
(FUN must take in a single argument and return a scalar)
if id.field is not NULL, AGG.FUN is a second fun to aggregate values from id.field to output interval
}

