% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.quantile}
\alias{gr.quantile}
\title{gr.quantile}
\usage{
gr.quantile(gr, qs, field = values(gr)[1], na.rm = FALSE)
}
\arguments{
\item{gr}{GRanges}

\item{qs}{quantiles ie numeric between 0 and 1
larger than 1, both boundary will be considered individual breakpoints}

\item{field}{metadata field of gr to query (default is the first metadata field)}
}
\value{
numeric vector of length qs corresponding to the width weighted quantiles specified by qs
}
\description{
Computes width weighted quantiles of some GRanges metadata field
"field"
}
\author{
Marcin Imielinski
}
