% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{\%Q\%}
\alias{\%Q\%}
\title{query ranges by applying an expression to ranges metadata}
\usage{
x \%Q\% ...
}
\value{
subset of gr that matches query
}
\description{
gr %Q% query returns the subsets of gr1 that matches meta data statement in query
}
\author{
Marcin Imielinski
}

