% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.rand}
\alias{gr.rand}
\title{Generate random \code{GRanges} on genome}
\usage{
gr.rand(w, genome)
}
\arguments{
\item{w}{Vector of widths (length of \code{w} determines length of output)}

\item{genome}{Genome which can be a \code{GRanges}, \code{GRangesList}, or \code{Seqinfo} object. Default is "hg19" from the \code{BSGenome} package.}
}
\value{
\code{GRanges} with random intervals on the specifed "chromosomes"
}
\description{
Randomly generates non-overlapping \code{GRanges} with supplied widths on supplied genome.
Seed can be supplied with \code{set.seed}
}
\note{
This function is currently quite slow, needs optimization
}
\examples{
## Generate 5 non-overlapping regions of width 10 on hg19
gr.rand(rep(10,5), BSgenome.Hsapiens.UCSC.hg19::Hsapiens)
}

