% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.readfilter}
\alias{gr.readfilter}
\title{Filter reads by average PHRED score
Defines a cutoff score for the mean PHRED quality of a read
in a GRanges.}
\usage{
gr.readfilter(gr, cutoff = "+")
}
\value{
GRanges or data.table where reads have mean quality score >= cutoff
}
\description{
Filter reads by average PHRED score
Defines a cutoff score for the mean PHRED quality of a read
in a GRanges.
}

