% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.refactor}
\alias{gr.refactor}
\title{gr.refactor}
\usage{
gr.refactor(gr, sn, gap = 0, rev = FALSE)
}
\arguments{
\item{gr}{\code{GRanges} to refactor}

\item{sn}{character vector of new seqnames}

\item{gap}{Default 0}

\item{rev}{Default FALSE}
}
\description{
Takes a pile of ranges gr and new seqnames "sn" (either of length 1 or
of length(gr)) and returns a gr object with the new seqnames and same
widths and new start coordinates.  These coordinates are determined by placing
each gr on the corresponding chromosome at 1 + gap after previous gr (or at 1)
}

