% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.resize}
\alias{gr.resize}
\title{Resize granges without running into negative width error}
\usage{
gr.resize(
  gr,
  width,
  pad = TRUE,
  minwid = 0,
  each = TRUE,
  ignore.strand = FALSE,
  fix = "center",
  reduce = FALSE
)
}
\value{
GRanges
}
\description{
lower size limit of window is 0
}
\author{
Kevin Hadi
}
