% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{\%_\%}
\alias{\%_\%}
\title{setdiff shortcut (strand agnostic)}
\usage{
x \%_\% ...
}
\value{
granges representing setdiff of input interval
}
\description{
Shortcut for setdiff

gr1 %_% gr2
}
\author{
Marcin Imielinski
}

