% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.setdiff2}
\alias{gr.setdiff2}
\title{gr.setdiff that works with multiple by columns}
\usage{
gr.setdiff2(query, subject, ignore.strand = TRUE, by = NULL, new = TRUE, ...)
}
\value{
GRanges
}
\description{

}
\author{
Kevin Hadi
}
