% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.simplify}
\alias{gr.simplify}
\title{Calculates pairwise distance for rearrangements represented by \code{GRangesList} objects}
\usage{
gr.simplify(gr, field = NULL, val = NULL, include.val = TRUE,
  split = FALSE, pad = 1)
}
\arguments{
\item{gr}{GRanges or GRangesList input}

\item{field}{character scalar, corresponding to value field of gr. (default = NULL)}

\item{val}{character scalar (default = NULL)}

\item{include.val}{boolean Flag will include in out gr values field of first matching record in input gr. (default = TRUE)}

\item{split}{boolean Flag to split the output into \code{GRangesList} split by \code{"field"}. (default = FALSE)}

\item{pad}{integer Pad ranges by this amount before doing merge. [1], which merges contiguous but non-overlapping ranges (default = 1)}
}
\value{
Simplified GRanges with "field" populated with uniquely contiguous values
}
\description{
Calculates pairwise distance for rearrangements represented by \code{GRangesList} objects
}
