% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.sort}
\alias{gr.sort}
\title{sort granges, grangeslist}
\usage{
gr.sort(gr, ignore.strand = TRUE)
}
\value{
GRanges
}
\description{
sort granges or grangeslist by seqlevels
also reorders seqlevels into 1:22, X, Y format
}
\author{
Kevin Hadi
}
