% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.splgaps}
\alias{gr.splgaps}
\title{gaps on GRanges, splitting by values in a metadata field}
\usage{
gr.splgaps(
  gr,
  ...,
  ignore.strand = TRUE,
  sep = paste0(" ", rand.string(length = 8), " "),
  start = 1L,
  end = seqlengths(gr),
  cleannm = TRUE,
  expand.levels = TRUE
)
}
\value{
GRangesList
}
\description{

}
\author{
Kevin Hadi
}
