% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.spreduce}
\alias{gr.spreduce}
\title{reduce based on a field(s) to split by in elementMetadata of GRanges, or given vector}
\usage{
gr.spreduce(
  gr,
  ...,
  ignore.strand = FALSE,
  pad = 0,
  return.grl = FALSE,
  sep = paste0(" ", rand.string(length = 8), " ")
)
}
\value{
GRanges
}
\description{
split and reduce GRanges by field(s)
if providing a variable not already within the GRanges,
may need to use dynget(variable_name)

Note: grl.reduce is slower than GenomicRanges::reduce(grangeslist)
}
\author{
Kevin Hadi
}
