% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.start}
\alias{gr.start}
\title{Get GRanges corresponding to beginning of range}
\usage{
gr.start(x, width = 1, force = FALSE, ignore.strand = TRUE,
  clip = FALSE)
}
\arguments{
\item{x}{\code{GRanges} object to operate on}

\item{width}{[default = 1] Specify subranges of greater width including the start of the range.}

\item{force}{[default = F] Allows returned \code{GRanges} to have ranges outside of its \code{Seqinfo} bounds.}

\item{ignore.strand}{If set to \code{FALSE}, will extend '-' strands from the other direction [TRUE].}

\item{clip}{[default = F] Trims returned \code{GRanges} so that it does not extend beyond bounds of the input \code{GRanges}}
}
\value{
\code{GRanges} object of width 1 ranges representing start of each genomic range in the input.
}
\description{
Get GRanges corresponding to beginning of range
}
\examples{
<<<<<<< HEAD
  \dontrun{st <- c(1,2,3)
  si <- Seqinfo("1",3)
  gr.start(GRanges(1, IRanges(st, width=101), seqinfo=si), width=20)}
=======
library(gUtils)
gr.start(gr.DNAase, width=200)
gr.start(gr.DNAase, width=200, clip=TRUE)
>>>>>>> 60083aa350ab9c4388ffd83f1d307cb2446af96b
}

