% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.string}
\alias{gr.string}
\title{Return UCSC style interval string corresponding to \code{GRanges} pile (ie chr:start-end)}
\usage{
gr.string(
  gr,
  add.chr = FALSE,
  mb = FALSE,
  round = 3,
  other.cols = c(),
  pretty = FALSE
)
}
\arguments{
\item{gr}{\code{GRanges} pile to get intervals from}

\item{add.chr}{boolean Flag to prepend seqnames with "chr" (default = FALSE)}

\item{mb}{boolean Flag to round to the nearest megabase (default = FALSE)}

\item{round}{integer If \code{mb} supplied, the number of digits to round to. (default = 3)}

\item{other.cols}{character vector Names of additional \code{mcols} fields to add to the string (seperated by ";")}

\item{pretty}{boolean Flag to output interval string in more readable format}
}
\value{
UCSC style interval string corresponding to \code{GRanges} pile
}
\description{
Return UCSC style interval string corresponding to \code{GRanges} pile (ie chr:start-end)
}
\examples{
gr.string(example_genes, other.cols = c("name", "name2"))
}
\author{
Marcin Imielinski
}
