% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.stripstrand}
\alias{gr.stripstrand}
\title{gr.stripstrand}
\usage{
gr.stripstrand(gr)
}
\arguments{
\item{gr}{\code{GRanges} to remove strand information from}
}
\value{
\code{GRanges} with strand set to \code{*}
}
\description{
Sets strand to "*"
}
