% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.sub}
\alias{gr.sub}
\title{Apply \code{gsub} to seqlevels of a \code{GRanges}}
\usage{
gr.sub(gr, a = c("^chr", "MT"), b = c("", "M"))
}
\arguments{
\item{gr}{\code{GRanges} to switch out seqlevels for}

\item{a}{vector of regular expressions of things to substitute out (default = c("^chr", "MT"))}

\item{b}{vector of values to substitute in (default = c("", "M"))}
}
\value{
GRanges with substitutions
}
\description{
Apply gsub to seqlevels of gr, by default removing 'chr', and "0.1" suffixes, and replacing "MT" with "M"
}
