% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.sum}
\alias{gr.sum}
\title{gr.sum}
\usage{
gr.sum(gr, field = NULL, mean = FALSE)
}
\arguments{
\item{gr}{\code{GRanges} to sum}

\item{field}{metadata field from gr to use as a weight}

\item{mean}{logical scalar specifying whether to divide the output at each interval but the total number of intervals overlapping it (only applies if field == NULL) (default FALSE)}
}
\value{
non-overlapping GRanges spanning the seqlengths of gr with $score (if field is NULL) or $field specifying the sum / mean at that position
}
\description{
Sums GRanges either by doing coverage and either weighting them equally
or using a field "weight".  Will return either sum or average.
}
