% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.tfix}
\alias{gr.tfix}
\title{gr.tfix}
\usage{
gr.tfix(gr, unmap = TRUE, readlen = 101)
}
\arguments{
\item{gr}{Input GRanges}

\item{unmap}{Toggle whether to turn start=1 reads onto the Unmapped seqlevel (and adds Unmapped). Default TRUE}
}
\description{
more aggressive version of gr.fix that requires seqlevels of c(1...22,X,Y,M)
and that the seqlengths agree with Hsapiens.
}

