% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gr.tile.map}
\alias{gr.tile.map}
\title{gr.tile.map}
\usage{
gr.tile.map(query, subject, mc.cores = 1, verbose = F)
}
\description{
Given two tilings of the genome (eg at different resolution)
query and subject outputs a length(query) list whose items are integer vectors of indices in subject
overlapping that overlap that query (strand non-specific)
}
\note{
Assumes that input query and subject have no gaps (including at end) or overlaps, i.e. ignores end()
coordinates and only uses "starts"
}

