% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.tile}
\alias{gr.tile}
\title{Tile ranges across a genomic range}
\usage{
gr.tile(gr, w = 1000)
}
\arguments{
\item{gr}{\code{GRanges}, \code{seqlengths} or \code{seqinfo} range to tile. If has \code{GRanges} has overlaps, will reduce first.}

\item{w}{Width of each tile}
}
\description{
tiles interval (or whole genome) with segments of <= specified width.  Returns strandless gr
"tiles".
}
\details{
input can be seqinfo object (in which case whole genome will be tiled);
if inputted grs overlap, will first reduce then tile.
}

