% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.tostring}
\alias{gr.tostring}
\title{gr.tostring}
\usage{
gr.tostring(gr, places = 2, interval = 1e+06, unit = "MB",
  prefix = "chr")
}
\description{
dumps out a quick text representation of a gr object (ie a character vector)
}

