% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.trim}
\alias{gr.trim}
\title{Trim a pile of GRanges}
\usage{
gr.trim(gr, starts = 1, ends = 1, fromEnd = FALSE, ignore.strand = T)

gr.trim(gr, starts = 1, ends = 1, fromEnd = FALSE, ignore.strand = T)
}
\arguments{
\item{gr}{GRanges to trime}

\item{starts}{number. Default 1}

\item{ends}{number. Default 1}

\item{gr}{GRanges to trime}

\item{starts}{number. Default 1}

\item{ends}{number. Default 1}
}
\description{
trims pile of granges relative to the specified <local> coordinates of each range
(ie the first coordinate of every gr is 1 and the last is width(gr))
if end is larger than the width of the corresponding gr, then the corresponding output will only have end(gr) as its coordinate.

trims pile of granges relative to the specified <local> coordinates of each range
(ie the first coordinate of every gr is 1 and the last is width(gr))
if end is larger than the width of the corresponding gr, then the corresponding output will only have end(gr) as its coordinate.
}
\details{
this is a role not currently provided by the standard granges fxns
(eg shift, reduce, restrict, shift, resize, flank etc)

this is a role not currently provided by the standard granges fxns
(eg shift, reduce, restrict, shift, resize, flank etc)
}

