% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.union.unstranded}
\alias{gr.union.stranded}
\alias{gr.union.unstranded}
\title{union shortcut}
\usage{
x \%|\% ...

x \%||\% ...
}
\value{
granges representing setdiff of input interval

granges representing setdiff of input interval
}
\description{
Shortcut for union

gr1 %|% gr2

Shortcut for setdiff

gr1 %||% gr2
}
\author{
Marcin Imielinski

Marcin Imielinski
}

