% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\docType{methods}
\name{\%OO\%}
\alias{\%OO\%}
\alias{\%OO\%,GRanges-method}
\title{gr.val shortcut to get fractional overlap of gr1 by gr2, respecting strand}
\usage{
x \%OO\% ...
}
\arguments{
\item{x}{See \link{gr.val}}

\item{...}{See \link{gr.val}}
}
\value{
fractional overlap  of gr1 with gr2
}
\description{
Shortcut for gr.val (using val = names(values(y)))

gr1 %OO% gr2
}
\author{
Marcin Imielinski
}

