% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{\%$$\%}
\alias{\%$$\%}
\alias{\%$$\%,GRanges-method}
\title{gr.val shortcut to get mean values of subject "x" meta data fields in query "y" (strand-specific)}
\usage{
x \%$$\% ...
}
\arguments{
\item{x}{See \link{gr.val}}

\item{...}{See \link{gr.val}}
}
\value{
gr1 with extra meta data fields populated from gr2
}
\description{
Shortcut for gr.val (using val = names(values(y)))

gr1 %$$% gr2
}
\author{
Marcin Imielinski
}
