% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{\%NN\%}
\alias{\%NN\%}
\alias{\%NN\%,GRanges-method}
\title{gr.val shortcut to get total numbers of intervals in gr2 overlapping with each interval in  gr1, respecting strand}
\usage{
x \%NN\% ...
}
\arguments{
\item{x}{See \link{gr.val}}

\item{...}{See \link{gr.val}}
}
\value{
bases overlap  of gr1 with gr2
}
\description{
gr1 %NN% gr2
}
\author{
Marcin Imielinski
}
