% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{\%oo\%}
\alias{\%oo\%}
\alias{\%oo\%,GRanges-method}
\title{gr.val shortcut to total per interval width of overlap of gr1 with gr2, respecting strand}
\usage{
x \%oo\% ...
}
\arguments{
\item{x}{See \link{gr.val}}

\item{...}{See \link{gr.val}}
}
\value{
bases overlap  of gr1 with gr2
}
\description{
Shortcut for gr.val (using val = names(values(y)))

gr1 %oo% gr2
}
\author{
Marcin Imielinski
}

