% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gr2circos}
\alias{gr2circos}
\title{gr2circos}
\usage{
gr2circos(outdir, ra = NULL, cn = NULL, cn.raw = NULL,
  data.field = "cn", text = NULL, text.field = "label", text.size = 24,
  text.r0 = 1.3, text.r1 = 1.5, ra.col = NULL,
  karyotype.fn = paste(CIRCOS.DIR, "data", "karyotype.hg19.txt", sep = "/"),
  png = T, svg = T, cn.low = 1, cn.high = 1.6, cn.line = FALSE,
  cn.cex = 1, cn.ylim = NULL, cn.tick = 1, cn.r0 = 1, cn.r1 = 1.3,
  chroms = NULL)
}
\description{
dumps rearrangemnts and cn data stored in GRanges into a simple circos format
}
\details{
ra = GRangesList output of ra_breaks
cn = outer ring of segment data
cn.raw = outer ring of scatter data
labels = GRanges of labels to draw in outer, outer ring

the plotted field in cn and cn.raw is specified by data.field
$col field of cn and cn.raw will specify color of segment / scatter point
$col field of text

if cn does not contain $col field then colors are automatically populated (blue for intervals with value below cn.low
and red for intervals with values above cn.high)

ra.col vector or $col field of ra (GRangesList value field) can specify colors (with ra.col taking precedence)
}

