% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr2dt}
\alias{gr2dt}
\title{Converts \code{GRanges} to \code{data.table}}
\usage{
gr2dt(gr)
}
\arguments{
\item{gr}{\code{GRanges} pile to convert to \code{data.table}}
}
\description{
Converts \code{GRanges} to \code{data.table}
}
\examples{
library(gUtils)
#gr2dt(GRanges("X", IRanges(1,2)))
}

