% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr2gatk}
\alias{gr2gatk}
\title{Dump GRanges to GATK file}
\usage{
gr2gatk(gr, file, add.chr = FALSE)
}
\arguments{
\item{gr}{GRanges}

\item{file}{file}

\item{add.chr}{Flag to add "chr" to seqnames. Default FALSE}
}
\value{
returns 0 if completed
}
\description{
Dumps gr object into gatk intervals in file path "file"
}

