% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr2grl}
\alias{gr2grl}
\title{gr2grl
Quick way to make grl from list of indices into a GRanges gr}
\usage{
gr2grl(gr, ix)
}
\arguments{
\item{gr}{\code{GRanges} to split}

\item{ix}{vector to split on}
}
\description{
gr2grl
Quick way to make grl from list of indices into a GRanges gr
}

