% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr_deconstruct_by}
\alias{gr_deconstruct_by}
\title{removing by field and random string barcode to seqnames for more efficient by queries}
\usage{
gr_deconstruct_by(x, by = NULL)
}
\arguments{
\item{by}{anything non-NULL to mark that by field was non-null in gr_construct_by}
}
\value{
A GRanges with the by metadata field attached to the seqnames
}
\description{
to be used with gr_construct_by
}
\author{
Kevin Hadi
}
