% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{\%--\%}
\alias{\%--\%}
\alias{\%--\%,}
\alias{GRangesList-method}
\title{Shift GRangesList left}
\usage{
grl \%--\% ...
}
\value{
shifted GRangesList
}
\description{
Operator to shift GRangesList left "sh" bases
}
