% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{grdt}
\alias{grdt}
\title{Converts \code{GRanges} to \code{data.table}}
\usage{
grdt(x)
}
\arguments{
\item{x}{\code{GRanges} to convert}
}
\description{
and a field grl.iix which saves the (local) index that that gr was in its corresponding grl item
}

