% Generated by roxygen2 (4.0.2): do not edit by hand
\name{grl.allin}
\alias{grl.allin}
\alias{grl.in}
\title{grl.allin}
\usage{
grl.in(grl, windows, some = F, only = F)
}
\description{
Like %in% for grl but now will return a logical vector that is true at position if i
only if the ranges in grl[i] intersect <<all>>, <<some>>, <<only>>  windows in the subject
}
\details{
eg can use to identify read pairs whose ends are contained inside two genes)
}

