% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{grl.bind}
\alias{grl.bind}
\title{Concatenate \code{GRangesList} objects.}
\usage{
grl.bind(...)
}
\arguments{
\item{...}{Any number of \code{GRangesList} to concatenate together}
}
\value{
Concatenated \code{GRangesList} with NA filled in for \code{mcols} fields that are non-overlapping. Note that the
elements are re-named with sequential numbers
}
\description{
Concatenates \code{GRangesList} objects taking the union of their \code{mcols} features if they have non-overlapping features
}
\examples{
## Concatenate
grl.hiC2 <- grl.hiC[1:20]
mcols(grl.hiC2)$test = 1
grl.bind(grl.hiC2, grl.hiC[1:30])
}
\author{
Marcin Imielinski
}
