% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{grl.end}
\alias{grl.end}
\title{Get the right ends of a \code{GRangesList}}
\usage{
grl.end(grl, width = 1, force = FALSE, ignore.strand = TRUE, clip = TRUE)
}
\arguments{
\item{width}{integer Specify subranges of greater width including the start of the range. (default = 1)}

\item{force}{boolean Allows returned \code{GRangesList} to have ranges outside of its \code{Seqinfo} bounds. (default = FALSE)}

\item{ignore.strand}{boolean If set to \code{FALSE}, will extend '-' strands from the other direction. (default = TRUE)}

\item{clip}{boolean Trims returned \code{GRangesList} so that it does not extend beyond bounds of the input (default = TRUE)}

\item{x}{\code{GRangesList} object containing GRanges to operate on}
}
\value{
GRangesList object of width = \code{width} ranges representing end of each genomic range in the input.
}
\description{
Function wrapping around gr.end to work on
GRangesLists
}
\author{
Kevin Hadi
}
