% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{grl.eval}
\alias{grl.eval}
\title{Evaluate and aggregate expression on GRanges column in GRangesList}
\usage{
grl.eval(grl, expr, condition = NULL)
}
\arguments{
\item{grl}{GRangesList to eval over}

\item{expr}{any syntactically valid R expression, on columns of GRanges or GRangesList}

\item{condition}{optional any syntactically valid R expression, on columns of GRanges or GRangesList, on which to subset before evaluating main 'expr'}
}
\description{
Evaluate expression 'expr' on indivdual GRanges inside GRangesList.
Expression should result in a single i.e. scalar value per grangeslist item.
}
