% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{++}
\alias{++}
\alias{++,GRangesList-method}
\title{Expand GRangesList left and right}
\usage{
"++"(grl, ...)
}
\value{
shifted GRangesList
}
\description{
Operator to shift GRangesList left and right "sh" bases
}
\author{
Kevin Hadi
}
