% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\docType{data}
\name{grl.hiC}
\alias{grl.hiC}
\title{HiC data for chr14 from Lieberman-Aiden 2009 (in hg19), subsampled
to 10,000 interactions}
\format{\code{GRangesList}}
\description{
HiC data for chr14 from Lieberman-Aiden 2009 (in hg19), subsampled
to 10,000 interactions
}
\keyword{data}
