% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\docType{methods}
\name{\%^^\%}
\alias{\%^^\%}
\alias{\%^\%,GRanges-method}
\title{grl.in shortcut}
\usage{
\S4method{\%^\%}{GRangesList}(x, y)
}
\arguments{
\item{x}{See \link{gr.in}}

\item{...}{See \link{gr.in}}
}
\value{
logical vector of length length(grl) which is TRUE at entry i only if grl[i] intersects at least one interval in gr (strand agnostic)
}
\description{
Shortcut for grl.in (standard arguments)

grl %^% gr
}
