% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{grl.in}
\alias{grl.in}
\title{Check intersection of \code{GRangesList} with windows on genome}
\usage{
grl.in(grl, windows, some = FALSE, only = FALSE, logical = FALSE, ...)
}
\arguments{
\item{grl}{\code{GRangesList} object to query for membership in \code{windows}}

\item{windows}{\code{GRanges} pile of windows}

\item{some}{Will return \code{TRUE} for \code{GRangesList} elements that intersect at least on window range [FALSE]}

\item{only}{Will return \code{TRUE} for \code{GRangesList} elements only if there are no elements of query that fail to interesect with windows [FALSE]}

\item{logical}{will return logical otherwise will return numeric vector of number of windows overlapping each grl}

\item{...}{Additional parameters to be passed on to \code{GenomicRanges::findOverlaps}}
}
\description{
Like %in% for grl but now will return a logical vector that is true at position if i
only if the ranges in grl[i] intersect <<all>>, <<some>>, <<only>>  windows in the subject
}
\details{
eg can use to identify read pairs whose ends are contained inside two genes)
}

