% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{grl.pivot}
\alias{grl.pivot}
\title{grl.pivot}
\usage{
grl.pivot(x)
}
\arguments{
\item{x}{GenomicRangesList object to pivot}
}
\description{
"pivots" grl object "x" by returning a new grl "y" whose
kth item is gr object of ranges x[[i]][k] for all i in 1:length(x)
}
\details{
Assumes all grs in "x" are of equal length
}

