% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\docType{methods}
\name{\%QQ\%}
\alias{\%QQ\%}
\alias{\%QQ\%,GRanges-method}
\title{query ranges by applying an expression to GRanges metadata within a GRangesList}
\usage{
x \%QQ\% ...
}
\arguments{
\item{x}{\code{GRangesList}}

\item{y}{expression querying metadata columns evaluating to a logical or integer matching to a subset of GRanges elements of x}
}
\value{
subset of grl that matches query
}
\description{
grl %QQ% query returns the subsets of GRanges elements within grl that matches meta data statement in query
}
\author{
Kevin Hadi
}
