% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{grl.reduce}
\alias{grl.reduce}
\title{grl.reduce}
\usage{
grl.reduce(grl, pad = 0, clip = FALSE)
}
\arguments{
\item{grl}{\code{GRangesList} input}

\item{pad}{integer Padding to add to ranges inside GRangesList before reducing (default = 0)}

\item{clip}{boolean Flag to add to ranges inside GRangesList before reducing (default = FALSE)}
}
\value{
\code{GRangesList} with GRanges of intervals "original GRanges +/- padding"
}
\description{
Quickly computes GRanges +/- padding inside a GRangesList
Can use with split / unlist
}
\author{
Marcin Imielinski
}
