% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{--}
\alias{--}
\alias{--,GRangesList-method}
\title{Shrink GRangesList left and right}
\usage{
"--"(grl, ...)
}
\value{
shifted GRangesList
}
\description{
Operator to shrink GRangesList left and right "sh" bases
}
\author{
Kevin Hadi
}
