% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{grl.shrink}
\alias{grl.shrink}
\title{grl.srhink}
\usage{
grl.shrink(grl, shrink_win)
}
\arguments{
\item{grl}{\code{GRangesList}}

\item{shrink_win}{\code{integer}}
}
\value{
GRangesList with shrunken windows
}
\description{
Function wrapping around the - operator
for GRanges objects to work on GRangesLists.
Shrnks window of element GRanges within GRangesList
}
