% Generated by roxygen2 (4.0.2): do not edit by hand
\name{grl.span}
\alias{grl.span}
\title{grl.span}
\usage{
grl.span(grl, chr = NULL, ir = F, keep.strand = T)
}
\description{
Returns GRanges object representing the left / right extent of each GRL item.  In case of "chimeric" GRL items (ie that map
to two chromosomes) there are two options:
(1) specify "chr" chromosome as argument to subset GRL's that are on that chromosome, and compute GRL extents from this, any GRL
    full outside of that chromosome will get a 0 width GRL
(2) (default) allow chimeric GRL items to get an extent that is with respect to the first chromosome in that GRL
}
\details{
If a grl item contains ranges that lie on different chromosomes, then corresponding grange will have chromosome "NA" and IRange(0, 0)
}

