% Generated by roxygen2 (4.0.2): do not edit by hand
\name{grl.split}
\alias{grl.split}
\title{grl.split}
\usage{
grl.split(grl, seqname = TRUE, strand = TRUE, values = c())
}
\description{
splits GRL's with respect to their seqnames and strand (default), returning
new grl whose items only contain ranges with a single seqname / strand
}
\details{
can also split by arbitrary (specified) genomic ranges value fields
}

