% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{grl.stripnames}
\alias{grl.stripnames}
\title{grl.stripnames}
\usage{
grl.stripnames(grl)
}
\arguments{
\item{grl}{\link{GRangesList} to string names from}
}
\description{
get rid of gr names inside a grl
}

