% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{grlbind}
\alias{grlbind}
\title{Concatenate \code{GRangesList} objects}
\usage{
grlbind(...)
}
\arguments{
\item{...}{Any number of \code{GRangesList} to concatenate together}
}
\value{
Concatenated GRangesList
}
\description{
Concatenates \code{GRangesList} objects taking the union of their \code{mcols} features if they have non-overlapping features
}
\examples{
## Concatenate
#grl.hiC2 <- grl.hiC[1:20]
#mcols(grl.hiC2)$test = 1
#grlbind(grl.hiC2, grl.hiC[1:30])
}

