% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gstring}
\alias{gstring}
\title{gstring}
\usage{
gstring(...)
}
\description{
quick function to parse gr from character vector IGV / UCSC style strings of format gr1;gr2;gr3 where each gr is of format chr:start-end[+/-]
}

