% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils-package.R
\docType{package}
\name{gUtils-package}
\alias{gUtils}
\alias{gUtils-package}
\title{gUtils - Extending GenomicRanges}
\description{
R package providing additional capabilities and speed for GenomicRanges operations.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{http://github.com/mskilab/gUtils/issues}
}

}
\author{
\strong{Maintainer}: Marcin Imielinski \email{mimielinski@nygenome.org}

Authors:
\itemize{
  \item Jeremiah Wala \email{jwala@broadinstitute.org}
}

}
