% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{hg_seqlengths}
\alias{hg_seqlengths}
\title{Output standard human genome seqlengths}
\usage{
hg_seqlengths(genome = NULL, chr = FALSE, include.junk = FALSE)
}
\arguments{
\item{genome}{A \code{BSgenome} or object with a \code{seqlengths} accessor. Default is hg19, but loads with warning unless explicitly provided}

\item{chr}{Flag for whether to keep "chr". Default FALSE}

\item{include.junk}{Flag for whether to not trim to only 1-22, X, Y, M. Default FALSE}
}
\value{
Named integer vector with elements corresponding to the genome seqlengths
}
\description{
Outputs a standard seqlengths for human genome +/- "chr".
}
\note{
A default genome can be set with the environment variable DEFAULT_BSGENOME. This
can be the full namespace of the genome  e.g.: \code{DEFAULT_BSGENOME=BSgenome.Hsapiens.UCSC.hg19::Hsapiens} OR  a URL / file path pointing to a chrom.sizes text file (e.g. http://genome.ucsc.edu/goldenpath/help/hg19.chrom.sizes) specifying a genome definition
}

