% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{import.ucsc}
\alias{import.ucsc}
\title{import.ucsc}
\usage{
import.ucsc(con, selection = NULL, text, chrsub = TRUE, verbose = FALSE,
  as = NULL, ...)
}
\description{
wrapper around rtracklayer import that
(1) handles "as" formats
(2) has additional flag chrsub to sub in 'chr' in selection, and then sub it out of the output
}

