% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{levapply}
\alias{levapply}
\title{levapply
Applies FUN locally to levels of x and returns vector of length()
(eg can do a "local" order within levels)}
\usage{
levapply(x, by, FUN = "order")
}
\arguments{
\item{x}{factor over which to apply the funciton}

\item{by}{group the factors?}

\item{FUN}{function to apply. Default \code{order}}
}
\description{
levapply
Applies FUN locally to levels of x and returns vector of length()
(eg can do a "local" order within levels)
}

