% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mafcount}
\alias{mafcount}
\title{mafcount}
\usage{
mafcount(tum.bam, norm.bam = NULL, maf, chunk.size = 100, verbose = TRUE,
  mc.cores = 1, ...)
}
\arguments{
\item{tum.bam}{path to a tumor bam file}

\item{norm.bam}{path to a normal bam file}

\item{maf}{path to a MAF}

\item{chunk.size}{Default 100}

\item{verbose}{Default TRUE}

\item{mc.cores}{Number of cores}

\item{...}{dummy}
}
\description{
Returns base counts for reference and alternative allele for an input tum and norm bam and maf data frame or \code{GRanges} specifying substitutions
}
\details{
maf is a single width GRanges describing variants and field \code{ref} (or \code{Reference_Allele}), \code{alt} (or \code{Tum_Seq_Allele1}) specifying reference and alt allele.
maf is assumed to have width 1 and strand is ignored.
}

