% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{match.bs}
\alias{match.bs}
\title{Identify matches between query and dictionary}
\usage{
match.bs(query, dict, midpoint = FALSE)
}
\arguments{
\item{query}{Query}

\item{dict}{Dictionary}

\item{midpoint}{Flag for output the coordinates of the match as the location,
where the midpoint of the dict string matches the given query. Default FALSE}
}
\value{
a vector of indices of length width(query) that contains
indices of the (starting) dictionary in the query string
}
\description{
Wrapper around matchPdict to identify matches between a query
string query and dictionary dict (both BString objects or subclasses)
}

