% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{mc.pairwiseAlignment}
\alias{mc.pairwiseAlignment}
\title{Multicore pairwiseAlignment}
\usage{
mc.pairwiseAlignment(pattern, subject, type = "local", mc.cores = 1,
  numchunk = NA, verbose = FALSE, ...)
}
\description{
Multicore implementation of pairwiseAlignment
}
\note{
WATCH OUT for bug where if any seqs sent to pairwiseAlignment start as empty, the
    C code underneath pairwiseAlignment crashes. Best to remove empties beforehand
}

