% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{munlist}
\alias{munlist}
\title{munlist}
\usage{
munlist(x, force.rbind = FALSE, force.cbind = FALSE, force.list = FALSE)
}
\description{
unlists a list of vectors, matrices, data frames into a n x k matrix
whose first column specifies the list item index of the entry
and second column specifies the sublist item index of the entry
and the remaining columns specifies the value(s) of the vector
or matrices.
}
\details{
force.cbind = TRUE will force concatenation via 'cbind'
force.rbind = TRUE will force concatenation via 'rxsbind'
}
\keyword{internal}

