% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{parse.gr}
\alias{parse.gr}
\title{parse.gr}
\usage{
parse.gr(...)
}
\arguments{
\item{...}{arguments to parse.grl i.e. character strings in UCSC style chr:start-end[+-]}
}
\value{
GRanges parsed from IGV-/UCSC-style strings
}
\description{
Quick function to parse GRanges from character vector IGV-/UCSC-style strings of format gr1;gr2;gr3 
where each input GRanges is of the format chr:start-end[+/-]
}
\author{
Marcin Imielinski
}
