% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{ra.overlaps}
\alias{ra.overlaps}
\title{Find overlaps between rearrangements represented by \code{GRangesList} objects}
\usage{
ra.overlaps(ra1, ra2, pad = 0, ...)
}
\arguments{
\item{ra1}{\code{GRangesList} of pairs of signed ranges representing a rearrangement set}

\item{ra2}{\code{GRangesList} of pairs of signed ranges representing a rearrangement set}

\item{pad}{Pad each breakpoint when considering overlaps. \code{[0]}}

\item{...}{Additional arguments to be sent to \code{\link{findOverlaps}} (e.g. \code{ignore.strand})}
}
\value{
\code{matrix} with the indices of \code{ra1} that overlap with \code{ra2} and vice-versa
}
\description{
Determines overlaps between two piles of rearrangement junctions, each \code{GRangesLists} of signed locus pairs,
against each other. returning a sparseMatrix that is T at entry ij if junction i overlaps junction j.
}

