% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{ra.stats}
\alias{ra.stats}
\title{Compare two rearrangement call sets}
\usage{
ra.stats(ra1, ra2, pad = 0, ignore.strand = FALSE)
}
\arguments{
\item{ra1}{\code{GRangesList} of rearrangement pairs}

\item{ra2}{\code{GRangesList} of rearrangement pairs}

\item{pad}{Amount to pad breakpoints by when considering overlaps}

\item{ignore.strand}{[default F] Should strandedness be considered when comparing calls}
}
\description{
Get union / interection counts and identities for two differrent
rearrangment call sets, as stored in \code{GRangesList} objects.
}
\seealso{
\code{\link{ra_breaks}} for importing dRanger breakpoint files,
\code{read_vcf} for reading in rearrangement VCF files.
}

