% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ra_breaks}
\alias{ra_breaks}
\title{ra_breaks}
\usage{
ra_breaks(rafile, keep.features = T, seqlengths = hg_seqlengths(),
  chr.convert = T, snowman = FALSE, breakpointer = FALSE,
  seqlevels = NULL, get.loose = FALSE)
}
\description{
takes in either file or data frame from dranger or snowman or path to BND / SV type vcf file
and returns junctions in VCF format.
}
\details{
The default output is GRangesList each with a length two GRanges whose strands point AWAY from the break.  If get.loose = TRUE (only relevant for VCF)
}

