% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{read.bam}
\alias{read.bam}
\title{Read BAM file into GRanges or data.table}
\usage{
read.bam(bam, intervals = NULL, gr = intervals, all = FALSE,
  bami = NULL, pairs.grl = TRUE, stripstrand = TRUE,
  what = scanBamWhat(), unpack.flag = FALSE, verbose = FALSE,
  tag = NULL, isPaired = NA, isProperPair = NA, isUnmappedQuery = NA,
  hasUnmappedMate = NA, isNotPassingQualityControls = NA, isDuplicate = F,
  isValidVendorRead = TRUE, as.grl = TRUE, as.data.table = FALSE,
  ignore.indels = FALSE, size.limit = 1e+06, ...)
}
\arguments{
\item{bam}{Input bam file. Advisable to make "bam" a BamFile instance instead of a plain string, so that the index does not have to be reloaded.}

\item{intervals}{GRanges of intervals to retrieve}

\item{gr}{GRanges of intervals to retrieve}

\item{bami}{Input bam index file.}

\item{stripstrand}{Flag to ignore strand information on the query intervals. Default TRUE}

\item{what}{What fields to pull down from BAM. Default \code{scanBamWhat()}}

\item{unpack.flag}{Add features corresponding to read flags. Default FALSE}

\item{verbose}{Increase verbosity}

\item{tag}{Additional tags to pull down from the BAM (e.g. 'R2')}

\item{isPaired}{See documentation for \code{scanBamFlag}. Default NA}

\item{isProperPair}{See documentation for \code{scanBamFlag}. Default NA}

\item{isUnmappedQuery}{See documentation for \code{scanBamFlag}. Default NA}

\item{hasUnmappedMate}{See documentation for \code{scanBamFlag}. Default NA}

\item{isNotPassingQualityControls}{See documentation for \code{scanBamFlag}. Default NA}

\item{isDuplicate}{See documentation for \code{scanBamFlag}. Default FALSE}

\item{isValidVendorRead}{See documentation for \code{scanBamFlag}. Default TRUE}

\item{as.grl}{Return reads as GRangesList. Controls whether \code{get.pairs.grl} does split. Default TRUE}

\item{as.data.table}{Return reads in the form of a data.table rather than GRanges/GRangesList}

\item{ignore.indels}{messes with cigar to read BAM with indels removed. Useful for breakpoint mapping on contigs}

\item{size.limit}{Default 1e6}

\item{...}{passed to \code{scanBamFlag}}
}
\value{
Reads in one of GRanges, GRangesList or data.table
}
\description{
Wrapper around Rsamtools bam scanning functions,
by default, returns GRangesList of read pairs for which <at least one> read lies in the supplied interval
}

