% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{read_hg}
\alias{read_hg}
\title{Wrapper around BSgenome call}
\usage{
read_hg(hg19 = T, fft = F)
}
\arguments{
\item{hg19}{Logical whether to return hg18 or hg19 BSgenome. Default TRUE}

\item{fft}{Logical whether to return an ffTrack. Default FALSE}
}
\value{
BSgenome or ffTrack of the genome
}
\description{
Retreives either the BSgenome hg18 or hg19 genome by default.  Requires packages
BSgenome.Hsapiens.UCSC.hg19 for hg19 and BSgenome.Hsapiens.UCSC.hg19 for hg18.
}
\details{
If fft = TRUE, can also also return the hg19 ffTrack (requires that the file exists)
Requires the existence of environment variable HG.FFT pointing to ffTrack .rds file..
}

