% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{rebin}
\alias{rebin}
\title{reaggregate WGS bins around a new target value}
\usage{
rebin(cov, binwidth, field = names(values(cov))[1], FUN = mean, na.rm = TRUE)
}
\arguments{
\item{cov}{GRanges of binned genome-wide coverage}

\item{binwidth}{new binwidth}
}
\value{
GRanges of binned genome-wide coverage at new bin
}
\description{
Given GRanges of bins will aggregate around new bin width.
}
\author{
Marcin Imielinski
}
