% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{remap_seqlevels}
\alias{remap_seqlevels}
\title{Remap seqlevels}
\usage{
remap_seqlevels(x, style = "NCBI")
}
\description{
Remap seqlevels using GenomeInfoDb
}
\details{
Wrapper to GenomeInfoDb functions to
remap seqlevels style. "NCBI" = no chr,
"UCSC" = chr prefixed.
}
\author{
Kevin Hadi
}
