% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{rle.query}
\alias{rle.query}
\title{Queries an \code{\link{RleList}} representing genomic data}
\usage{
rle.query(
  subject.rle,
  query.gr,
  chunksize = 1e+09,
  mc.cores = 1,
  verbose = FALSE
)
}
\arguments{
\item{subject.rle}{\code{Rle}}

\item{query.gr}{integer GRangeslist or GRanges}

\item{chunksize}{integer Number of \code{query.gr} ranges to consider in one memory chunk. (default = 1e9)}

\item{mc.cores}{Number of cores to apply when doing chunked operation (default = 1)}

\item{verbose}{boolean Set the verbosity of the output (default = FALSE)}
}
\value{
Rle representing the (concatenated) vector of data (reversing order in case of negative strand input)
}
\description{
Queries an \code{\link{RleList}} representing genomic data

(i.e. a list whose names represent seqnames ie chromosomes, and lengths represent seqlengths)
via \code{GRanges} object
}
\note{
Throws warning if seqlengths(gr) do not correspond to the lengths of the \code{RleList} components
}
