% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rrbind}
\alias{rrbind}
\title{rrbind = function(df1, df2, [df3 ... etc], )}
\usage{
rrbind(..., union = T)
}
\description{
like rbind, but takes the intersecting columns of the dfs
}
\details{
if union flag is used then will take union of columns (and put NA's for columns of df1 not in df2 and vice versa)
}

