% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{rrbind2}
\alias{rrbind2}
\title{Improved rbind for intersecting columns of data.frames or data.tables}
\usage{
rrbind2(..., union = TRUE, as.data.table = FALSE)
}
\arguments{
\item{union}{if union flag is used then will take union of columns (and put NA's for columns of df1 not in df2 and vice versa). Default TRUE}
}
\description{
like rbind, but takes the intersecting columns of the dfs
rrbind = function(df1, df2, [df3 ... etc], )
}
\keyword{internal}

