% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{seg2gr}
\alias{seg2gr}
\title{Convert GRanges-like data.frame into GRanges}
\usage{
seg2gr(segs, seqlengths = NULL, seqinfo = Seqinfo())
}
\arguments{
\item{segs}{data.frame (or data.table) of segments with fields denoting chromosome, start, end, and other metadata. (See function 'standardize_segs()' for 'seg' data.frame/data.table input formats)}

\item{seqlengths}{seqlengths of output GRanges object (default = NULL)}

\item{seqinfo}{seqinfo of output GRanges object (default = Seqinfo())}
}
\value{
GRanges from converted "segs" data.frame/data.table
}
\description{
Input data.frame of segments "segs" and converts into GRanges object porting over additional value columns

"segs" data.frame/data.table can obey any number of conventions to specify chrom, start, and end of ranges
(e.g. $pos1, $pos2, $Start_position, $End_position)

Please see documentation for function 'standardize_segs()' for more details.
}
