% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{seqinfo2gr}
\alias{seqinfo2gr}
\title{Create GRanges from Seqinfo}
\usage{
seqinfo2gr(si, strip.empty = FALSE)
}
\arguments{
\item{si}{Seqinfo object}

\item{strip.empty}{Don't know. Default FALSE}
}
\description{
Creates a genomic ranges from seqinfo object
ie a pile of ranges spanning the genome
}
\examples{
\dontrun{si <- Seqinfo(names(hg_seqlength(), hg_seqlengths()))
seqinfo2gr(si)}
}

