% Generated by roxygen2 (4.0.2): do not edit by hand
\name{si2gr}
\alias{si2gr}
\title{Create GRanges from Seqinfo}
\usage{
si2gr(si, strip.empty = FALSE)
}
\arguments{
\item{si}{Seqinfo object}

\item{strip.empty}{Don't know. Default FALSE}
}
\description{
Creates a genomic ranges from seqinfo object
ie a pile of ranges spanning the genome
}
\examples{
\dontrun{si <- Seqinfo(names(hg_seqlength(), hg_seqlengths()))
si2gr(si)}
}

