% Generated by roxygen2 (4.0.2): do not edit by hand
\name{splice.cigar}
\alias{splice.cigar}
\title{splice.cigar}
\usage{
splice.cigar(reads, verbose = T, fast = TRUE, use.D = TRUE,
  rem.soft = T, get.seq = FALSE, return.grl = TRUE)
}
\description{
takes gr or gappedalignment object "reads" and parses cigar fields
to return grl corresponding to spliced alignments on the genome corresponding to
portions of the cigar
}
\details{
ie each outputted grl item contains the granges corresponding to all non-N portions of cigar string

if grl provided as input (e.g. paired reads) then all of the spliced ranges resulting from each
input grl item will be put into the corresponding output grl item

NOTE: does not update MD tag

if use.D = TRUE, then will treat "D" (deletion) in addition to "N" flags as indicative of deletion event.
}

