% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{standardize_segs}
\alias{standardize_segs}
\title{Takes and returns segs data frame standardized to a single format}
\usage{
standardize_segs(seg, chr = FALSE)
}
\arguments{
\item{segs}{data frame of segments with fields denoting chromosome, start, end, and other metadata (see standardized segs for seg data frame input formats)}

\item{seqlengths}{seqlengths of output GRanges object}

\item{seqinfo}{seqinfo of output GRanges object}
}
\description{
(data frame seg function)

Takes and returns segs data frame standardized to a single format (ie $chr, $pos1, $pos2)

if chr = TRUE will ensure "chr" prefix is added to chromossome(if does not exist)
}
