% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{system.call}
\alias{system.call}
\title{Wrapper to base \code{system} function to call system (e.g. perl) from R.
The only benefit to this wrapper is a more controlled verbose argument.}
\usage{
system.call(syscall, verbose = T)
}
\arguments{
\item{syscall}{string containing the system call}

\item{verbose}{print the syscall to screen, and it's stdout}
}
\description{
Wrapper to base \code{system} function to call system (e.g. perl) from R.
The only benefit to this wrapper is a more controlled verbose argument.
}
\examples{
# system.call('perl s/[0-9]+//g file1 > file2')
}
\author{
Jeremiah Wala \email{jwala@broadinstitute.org}
}

