% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{vaggregate}
\alias{vaggregate}
\title{vaggregate}
\usage{
vaggregate(...)
}
\arguments{
\item{...}{things to aggregate}
}
\description{
same as aggregate except returns named vector
with names as first column of output and values as second
}

