% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{alpha}
\alias{alpha}
\title{alpha}
\usage{
alpha(col, alpha)
}
\arguments{
\item{col}{color string}

\item{alpha}{alpha value between 0 and 1}
}
\description{
takes provided colors and gives them the specified alpha (ie transparency) value
}

