% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{bamflag}
\alias{bamflag}
\alias{bamtag}
\title{bamflag}
\usage{
bamflag(reads)

bamtag(reads, secondary = F, gr.string = F)
}
\arguments{
\item{reads}{GenomicRanges holding the reads}

\item{reads}{GenomicRanges holding the reads}
}
\description{
shortcut .. assumes reads are GappedAlignments with flag variable or actual integers representing bam flag

outputs a tag that cats qname, first vs first second mate +/- secondary alignment +/- gr.string
to give an identifier for determine duplicates in a read pile
}

