% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bamtag}
\alias{bamtag}
\title{bamtag}
\usage{
bamtag(reads, secondary = F, gr.string = F)
}
\description{
outputs a tag that cats qname, first vs first second mate +/- secondary alignment +/- gr.string
to give an identifier for determine duplicates in a read pile
}

