% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{concatPA}
\alias{concatPA}
\title{Concatenate pairwiseAlignmen objects}
\usage{
concatPA(pa, pattern, subject)
}
\arguments{
\item{pa}{list of pairwiseAlignment objects}

\item{pattern}{Pattern of the entire thing}

\item{subject}{Subject of the entire thing}
}
\value{
A single \code{pairwiseAlignment} object
}
\description{
Concatenate pairwiseAlignment objects. modeled after \code{mpi.collate.pairwiseAlignment}
}

