% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{counts2rpkm}
\alias{counts2rpkm}
\title{Compute rpkm counts from counts}
\usage{
counts2rpkm(counts, by)
}
\arguments{
\item{counts}{GRanges, data.table or data.frame with records, width fields}

\item{by}{Field to group counts by}
}
\description{
takes countbam (or bam.cov.gr) output "counts" and computes rpkm by aggregating across "by" variable
}
\note{
The denominator (ie total reads) is just the sum of counts$records
}

