% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{dt2gr}
\alias{dt2gr}
\title{Convert data.table to GRanges}
\usage{
dt2gr(dt, key = NULL, seqlengths = hg_seqlengths(), seqinfo = Seqinfo())
}
\arguments{
\item{dt}{\code{data.table} to convert to \code{GRanges}}

\item{dt}{data.table to convert}
}
\value{
\code{GRanges} object of \code{length = nrow(dt)}
}
\description{
Takes as input a data.table which must have the following fields: \code{start}, \code{end}, \code{strand}, \code{seqnames}. Will throw
an error if any one of these is not present.
All of the remaining fields are added as metadata to the \code{GRanges}.
}
\examples{
gr <- dt2gr(data.table(start=c(1,2), seqnames=c("X", "1"), end=c(10,20), strand = c('+', '-')))
}

