% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{dtgr}
\alias{dtgr}
\title{Convert data.table to GRanges}
\usage{
dtgr(dt)
}
\arguments{
\item{dt}{data.table to convert to GRanges}
}
\value{
GRanges object of length = nrow(dt)
}
\description{
Takes as input a data.table which must have the fields: start, end, strand, seqnames.
All of the remaining fields are added as meta data to the GRanges
}
\examples{
\dontrun{r <- dtgr(data.table(start=1, seqnames="X", end=2, strand='+'))}
}

