% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gc_content}
\alias{gc_content}
\title{Get GC content from reference genome}
\usage{
gc_content(segs, bs_genome)
}
\arguments{
\item{segs}{Segment data frame to pull gc from}

\item{bs_genome}{A \code{\link{BSgenome}} object. Perhaps \code{BSgenome.Hsapiens.UCSC.hg19::Hsapiens}}
}
\description{
Uses BSgenome package to compute gc content for a collection of segments in seg data frame ($chr, $start, $end or $chr, $pos1, $pos2 or $chr, $begin, $end)
Returns vector of gc content of length nrow(segs).
}

