% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.chr}
\alias{gr.chr}
\title{Prepend "chr" to \code{GRanges seqlevels}}
\usage{
gr.chr(gr)
}
\arguments{
\item{gr}{\code{GRanges} object to append 'chr' to}
}
\value{
Identical \code{GRanges}, but with 'chr' prepended to each seqlevel
}
\description{
Prepend "chr" to \code{GRanges seqlevels}
}
\details{
implementation looks for strings that begin with 0-9, X,Y,M, or Un for
the unmapped chromosome regions/decoys.  
regex pattern: "^([0-9,X, Y, M,x,y,m, Un, {EBV}]+)"
replacement pattern: "chr\\1"
}
\examples{

gr <- gr.chr(GRanges(c(1,"chrX"), IRanges(c(1,2), 1)))
seqnames(gr)

}
\author{
Max Chao
}
