% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.chr}
\alias{gr.chr}
\title{Add "chr" to GRanges seqlevels}
\usage{
gr.chr(gr)
}
\arguments{
\item{gr}{GRanges object to append 'chr' to}
}
\value{
Identical GRanges, but with 'chr' appended to each seqlevel
}
\description{
Adds "chr" to seqlevels of gr / grl object to make compatible with UCSC genome
}

