% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.chr}
\alias{gr.chr}
\title{Prepend "chr" to \code{GRanges seqlevels}}
\usage{
gr.chr(gr)
}
\arguments{
\item{gr}{\code{GRanges} object to append 'chr' to}
}
\value{
Identical \code{GRanges}, but with 'chr' prepended to each seqlevel
}
\description{
Prepend "chr" to \code{GRanges seqlevels}
}
\examples{

gr <-  gr.chr(GRanges(c(1,"chrX"), IRanges(c(1,2), 1)))
seqnames(gr)

}
