% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.collapse}
\alias{gr.collapse}
\title{Collapse adjacent ranges}
\usage{
gr.collapse(gr, pad = 1)
}
\arguments{
\item{gr}{\code{GRanges} to collapse}

\item{pad}{Padding that allows for not quite adjacent elements to be considered overlapping. 1}
}
\value{
Collapsed ranges
}
\description{
Like \code{GenomicRanges::reduce} except only collapses <<adjacent>> ranges in the input
}

