% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.collapse}
\alias{gr.collapse}
\title{Collapse adjacent ranges}
\usage{
gr.collapse(gr, pad = 1)
}
\arguments{
\item{gr}{\code{GRanges} to collapse}

\item{pad}{Padding that allows for not quite adjacent elements to be considered overlapping. (default = 1)}
}
\value{
GRanges with collapsed adjacent GRanges
}
\description{
Like \code{GenomicRanges::reduce} except only collapses <<adjacent>> ranges in the input
}
\author{
Marcin Imielinski
}
