% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.collapse}
\alias{gr.collapse}
\title{gr.collapse}
\usage{
gr.collapse(gr, pad = 1)
}
\description{
like "reduce" except only collapses <<adjacent>> ranges in the input
returning the collapsed ranges
}

