% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.dist}
\alias{gr.dist}
\title{Pairwise distance between two \code{GRanges}}
\usage{
gr.dist(gr1, gr2 = NULL, ignore.strand = FALSE, ...)
}
\arguments{
\item{gr1}{First \code{GRanges}}

\item{gr2}{Second \code{GRanges}}

\item{ignore.strand}{Don't required elements be on same strand to avoid NA [FALSE]}

\item{...}{Additional arguments to be supplied to \code{GenomicRanges::distance}}
}
\value{
Matrix with the pairwise distances, with \code{gr1} on rows and \code{gr2} on cols
}
\description{
computes pairwise distance between elements of two gr objects of length n and m
returning n by m matrix of distances between item i of gr1 and item j of gr2
}
\details{
distances are computed as follows:
NA for ranges on different seqnames
0 for overlapping ranges)
min(abs(end1-end2), abs(end1-start2), abs(start1-end2), abs(start1-end1),) for all others

if only gr1 is provided, then will return n x n matrix of gr's vs themselves

if max.dist = TRUE then will replace min with max above
}

