% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.dist}
\alias{gr.dist}
\title{Pairwise distance between two \code{GRanges}}
\usage{
gr.dist(gr1, gr2 = NULL, ignore.strand = FALSE, ...)
}
\arguments{
\item{gr1}{First \code{GRanges}}

\item{gr2}{Second \code{GRanges}}

\item{ignore.strand}{Don't required elements be on same strand to avoid \code{NA [FALSE]}}

\item{...}{Additional arguments to be supplied to \code{GenomicRanges::distance}}
}
\value{
\code{N} by \code{M} matrix with the pairwise distances, with \code{gr1} on rows and \code{gr2} on cols
}
\description{
Computes matrix of pairwise distance between elements of two \code{GRanges} objects of length \code{n} and \code{m}.
}
\details{
Distances are computed as follows:
\itemize{
\item NA for ranges on different seqnames
\item 0 for overlapping ranges
\item min(abs(end1-end2), abs(end1-start2), abs(start1-end2), abs(start1-end1),) for all others
}
If only \code{gr1} is provided, then will return n x n matrix of \code{gr1} vs itself \cr
If \code{max.dist = TRUE}, then will replace \code{min} with \code{max} above
}

