% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\docType{data}
\name{gr.DNAase}
\alias{gr.DNAase}
\title{DNAaseI hypersensitivity sites from UCSC Table Browser hg19,
subsampled to 10,000 sites}
\format{\code{GRanges}}
\description{
DNAaseI hypersensitivity sites from UCSC Table Browser hg19,
subsampled to 10,000 sites
}
\keyword{data}

