% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.duplicated}
\alias{gr.duplicated}
\title{Allows to restrict duplicates using "by" columns and allows in exact matching}
\usage{
gr.duplicated(query, by = NULL, type = "any")
}
\arguments{
\item{query}{query ranges}
}
\description{
Allows to restrict duplicates using "by" columns and allows in exact matching
}
