% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.end}
\alias{gr.end}
\alias{\%)\%}
\alias{\%)\%,GRanges-method}
\alias{\%))\%}
\alias{\%))\%,GRanges-method}
\alias{\%(\%}
\alias{\%(\%,GRanges-method}
\alias{\%((\%}
\alias{\%((\%,GRanges-method}
\title{Get the right ends of a \code{GRanges}}
\usage{
gr.end(x, width = 1, force = FALSE, ignore.strand = TRUE, clip = TRUE)

x \%)\% ...

x \%))\% ...

x \%(\% ...

x \%((\% ...
}
\arguments{
\item{x}{\code{GRanges} object}

\item{width}{integer Specify subranges of greater width including the start of the range. (default = 1)}

\item{force}{boolean Allows returned \code{GRanges} to have ranges outside of its \code{Seqinfo} bounds. (default = FALSE)}

\item{ignore.strand}{boolean If set to \code{FALSE}, will extend '-' strands from the other direction. (default = TRUE)}

\item{clip}{boolean Trims returned \code{GRanges} so that it does not extend beyond bounds of the input (default = TRUE)}

\item{y}{num bases to get from left side of interval}
}
\value{
GRanges object of width = \code{width} ranges representing end of each genomic range in the input.

\code{GRanges} right sides of input intervals

\code{GRanges} 3' sides of input intervals

\code{GRanges} left sides of input intervals

\code{GRanges} 5' sides of input intervals
}
\description{
Alternative to \code{GenomicRanges::flank} that will provide end positions *within* intervals

Shortcut for \code{gr.end}

gr1 <- GRanges(1, IRanges(10,20), strand="+")
gr1 %)% 5 # right 5 bases of gr1

Shortcut for \code{gr.end}

gr1 <- GRanges(1, IRanges(10,20), strand="+")
gr1 %))% 5 # get three prime 5 bases of gr1

Shortcut for \code{gr.end}

gr1 <- GRanges(1, IRanges(10,20), strand="+")
gr1 %(% 5 # left 5 bases of gr1

Shortcut for \code{gr.end}

gr1 <- GRanges(1, IRanges(10,20), strand="+")
gr1 %((% 5 # get three prime 5 bases of gr1
}
\examples{
gr.end(example_dnase, width=200, clip=TRUE)
}
\author{
Marcin Imielinski

Marcin Imielinski

Marcin Imielinski

Marcin Imielinski

Marcin Imielinski
}
