% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.end}
\alias{gr.end}
\title{Get the right ends of a \code{GRanges}}
\usage{
gr.end(x, width = 1, force = FALSE, ignore.strand = TRUE, clip = TRUE)
}
\arguments{
\item{x}{\code{GRanges} object to operate on}

\item{width}{Specify subranges of greater width including the start of the range. \code{[1]}}

\item{force}{Allows returned \code{GRanges} to have ranges outside of its \code{Seqinfo} bounds. \code{[FALSE]}}

\item{ignore.strand}{If set to \code{FALSE}, will extend '-' strands from the other direction. \code{[TRUE]}}

\item{clip}{Trims returned \code{GRanges} so that it does not extend beyond bounds of the input \code{GRanges}. \code{[TRUE]}}
}
\value{
\code{GRanges} object of width = \code{width} ranges representing end of each genomic range in the input.
}
\description{
Alternative to \code{GenomicRanges::flank} that will provide end positions *within* intervals
}
\examples{
gr.end(example_dnase, width=200, clip=TRUE)
}

