% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.end}
\alias{gr.end}
\title{Get GRanges corresponding to beginning of end}
\usage{
gr.end(x, width = 1, force = F, ignore.strand = T, clip = T)
}
\arguments{
\item{x}{\code{GRanges} object to operate on}

\item{width}{[default = 1] Specify subranges of greater width including the start of the range.}

\item{force}{[default = F] Allows returned \code{GRanges} to have ranges outside of its \code{Seqinfo} bounds.}

\item{ignore.strand}{[default = T] If set to \code{FALSE}, will extend '-' strands from the other direction.}

\item{clip}{[default = F] Trims returned \code{GRanges} so that it does not extend beyond bounds of the input \code{GRanges}}
}
\value{
GRanges object of width 1 ranges representing end of each genomic range in the input.
}
\description{
Alternative to \code{flank} that will provide end positions *within* intervals
}
\examples{
gr.end(GRanges(1, IRanges(start=c(1,2,3), width=101), seqinfo=Seqinfo("1", 3)), width=20)
gr.end(GRanges(1, IRanges(start=c(1,2,3), width=101), seqinfo=Seqinfo("1", 3)), width=200, clip=T)
}

