% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.findoverlaps}
\alias{gr.findoverlaps}
\title{Faster replacement for \code{GRanges} version of \code{findOverlaps}}
\usage{
gr.findoverlaps(query, subject, ignore.strand = TRUE, first = FALSE,
  qcol = NULL, scol = NULL, max.chunk = 1e+13,
  foverlaps = ifelse(is.na(as.logical(Sys.getenv("GRFO_FOVERLAPS"))), FALSE,
  as.logical(Sys.getenv("GRFO_FOVERLAPS"))) & exists("foverlaps"),
  pintersect = NA, verbose = FALSE, type = "any", by = NULL,
  mc.cores = 1, return.type = "same", ...)
}
\arguments{
\item{query}{Query \code{GRanges} pile}

\item{subject}{Subject \code{GRanges} pile}

\item{ignore.strand}{Don't consider strand information during overlaps [TRUE]}

\item{first}{TODO}

\item{qcol}{\code{character} vector of query meta-data columns to add to results}

\item{scol}{\code{character} vector of subject meta-data columns to add to results}

\item{max.chunk}{Maximum number of ranges to consider in one chunk (to keep down memory) [1e13]}

\item{foverlaps}{Use \code{data.table::foverlaps} instead of \code{IRanges::findOverlaps}}

\item{pintersect}{TODO}

\item{verbose}{Increase the verbosity during runtime [FALSE]}

\item{type}{TODO}

\item{by}{Meta-data column to consider when performing overlaps [NULL]}

\item{mc.cores}{Number of cores to use, if ranges exceed \code{max.chunk} [1]}

\item{return.type}{TODO}

\item{...}{TODO}
}
\value{
\code{GRanges} pile of the intersection regions, with \code{query.id} and \code{subject.id} marking sources
}
\description{
returns granges of matches with two additional fields
$query.id - index of matching query
$subject.id - index of matching subject
}
\details{
pintersect employs pintersect to find overlaps, in general this is slower, but can be much faster with much lower
memory footprint for large ranges sets with many different seqnames (eg transcriptome)
max.chunk controls the maximum number of range pairs that compared at once

Optional "by" field is a character scalar that specifies a metadata column present in both query and subject
that will be used to additionally restrict matches, i.e. to pairs of ranges that overlap and also
have the same values of their "by" fields

... = additional args for \code{findOverlaps} (IRanges version)
}

