% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.findoverlaps}
\alias{gr.findoverlaps}
\title{Wrapper to \code{GenomicRanges::findOverlaps} with added functionality}
\usage{
gr.findoverlaps(query, subject, ignore.strand = TRUE, first = FALSE,
  qcol = NULL, scol = NULL, type = "any", by = NULL,
  return.type = "same", max.chunk = 10000000000000, verbose = FALSE,
  mc.cores = 1, ...)
}
\arguments{
\item{query}{Query \code{GRanges} pile}

\item{subject}{Subject \code{GRanges} pile}

\item{ignore.strand}{Don't consider strand information during overlaps. \code{[TRUE]}}

\item{first}{Restrict to only the first match of the subject (default is to return all matches). \code{[FALSE]}}

\item{qcol}{\code{character} vector of query meta-data columns to add to results}

\item{scol}{\code{character} vector of subject meta-data columns to add to results}

\item{type}{\code{type} argument as defined by \code{IRanges::findOverlaps} (\code{"any"}, \code{"start"}, \code{"end"}, \code{"within"}, \code{"equal"}). \code{["any"]}}

\item{by}{Meta-data column to consider when performing overlaps [NULL]}

\item{return.type}{Select data format to return (supplied as character): \code{"same"}, \code{"data.table"}, \code{"GRanges"}. \code{["same"]}}

\item{max.chunk}{Maximum number of \code{query*subject} ranges to consider at once. Lower number increases runtime but decreased memory. If \code{length(query)*length(subject)} is less than \code{max.chunk}, overlaps will run in one batch.\code{[1e13]}}

\item{verbose}{Increase the verbosity. \code{[FALSE]}}

\item{mc.cores}{Number of cores to use when running in chunked mode}

\item{...}{Additional arguments sent to \code{IRanges::findOverlaps}.}
}
\value{
\code{GRanges} pile of the intersection regions, with \code{query.id} and \code{subject.id} marking sources
}
\description{
Returns \code{GRanges} of matches with two additional fields:
\itemize{
\code{$query.id} - index of matching query
\code{$subject.id} - index of matching subject
}
Optional \code{"by"} field is a character scalar that specifies a metadata column present in both query and subject
that will be used to additionally restrict matches, i.e. to pairs of ranges that overlap and also
have the same values of their \code{"by"} fields
}

