% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.findoverlaps}
\alias{gr.findoverlaps}
\title{Wrapper to \code{GenomicRanges::findOverlaps} with added functionality}
\usage{
gr.findoverlaps(
  query,
  subject,
  ignore.strand = TRUE,
  first = FALSE,
  qcol = NULL,
  scol = NULL,
  type = "any",
  by = NULL,
  return.type = "same",
  max.chunk = 1e+13,
  verbose = FALSE,
  mc.cores = 1,
  maxgap = -1L,
  new.by = TRUE,
  ...
)
}
\arguments{
\item{query}{Query \code{GRanges} pile}

\item{subject}{Subject \code{GRanges} pile}

\item{ignore.strand}{boolean Strand ignored (default = TRUE)}

\item{first}{boolean Flag if TRUE restricts to only the first match of the subject. If FALSE will return all matches. (default = FALSE)}

\item{qcol}{\code{character} vector of query meta-data columns to add to results (default = NULL)}

\item{scol}{\code{character} vector of subject meta-data columns to add to results (default = NULL)}

\item{type}{\code{type} argument as defined by \code{IRanges::findOverlaps} (\code{"any"}, \code{"start"}, \code{"end"}, \code{"within"}, \code{"equal"}). (default = 'any')}

\item{by}{vector Meta-data column to consider when performing overlaps (default = NULL)}

\item{return.type}{character Select data format to return (supplied as character): \code{"same"}, \code{"data.table"}, \code{"GRanges"}. (default = 'same')}

\item{max.chunk}{integer Maximum number of \code{query*subject} ranges to consider at once. Lower number increases runtime but decreased memory. If \code{length(query)*length(subject)} is less than \code{max.chunk}, overlaps will run in one batch. (default = 1e3)}

\item{verbose}{boolean Flag to increase the verbosity. (default = FALSE)}

\item{mc.cores}{Number of cores to use when running in chunked mode (default = 1)}

\item{...}{Additional arguments sent to \code{IRanges::findOverlaps}.}
}
\value{
\code{GRanges} pile of the intersection regions, with \code{query.id} and \code{subject.id} marking sources
}
\description{
Wrapper to \code{GenomicRanges::findOverlaps} with added functionality

Returns \code{GRanges} of matches with two additional fields:
\itemize{
\item \code{$query.id} - index of matching query
\item \code{$subject.id} - index of matching subject
}
Optional \code{"by"} field is a character scalar that specifies a metadata column present in both query and subject
that will be used to additionally restrict matches, i.e. to pairs of ranges that overlap and also
have the same values of their \code{"by"} fields
}
