% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.fix}
\alias{gr.fix}
\title{"Fixes" \code{seqlengths} / \code{seqlevels}}
\usage{
gr.fix(gr, genome = NULL, gname = NULL, drop = FALSE)
}
\arguments{
\item{gr}{\code{GRanges} object to fix}

\item{genome}{Genome to fix to: \code{Seqinfo}, \code{BSgenome}, \code{GRanges} (w/seqlengths), \code{GRangesList} (w/seqlengths)}

\item{gname}{Name of the genome (optional, just appends to \code{Seqinfo} of the output) [NULL]}

\item{drop}{Remove ranges that are not present in the supplied genome [FALSE]}
}
\value{
\code{GRanges} pile with the fixed \code{Seqinfo}
}
\description{
If "genome" not specified will replace \code{NA} \code{seqlengths} in GRanges to reflect largest coordinate per \code{seqlevel}
and removes all \code{NA seqlevels} after this fix.
}
\details{
if "genome" defined (i.e. as \code{Seqinfo} object, or a \code{BSgenome}, \code{GRanges}, \code{GRangesList} object with populated \code{seqlengths}),
then will replace \code{seqlengths} in \code{gr} with those for that genome
}

