% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.fix}
\alias{gr.fix}
\title{gr.fix}
\usage{
gr.fix(gr, genome = NULL, gname = NULL, drop = FALSE)
}
\arguments{
\item{gr}{\code{GRanges} to be fixed}

\item{genome}{Genome to fix to}

\item{gname}{dummy}

\item{drop}{Remove all ranges without seqlevelts in the genome. Default FALSE}
}
\description{
"Fixes" seqlengths / seqlevels
If "genome" not specified will replace NA seq lengths in GR to reflect largest coordinate per seqlevel
and removes all NA seqlevels after this fix.
}
\details{
if "genome" defined (ie as Seqinfo object, or a BSgenome, GRanges, GRnagesList object with populated seqlengths) then will replace
seqlengths in gr with those for that genome (and if drop = T, drop all ranges without
seqlevels in that genome)
}

