% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.flatten}
\alias{gr.flatten}
\title{gr.flatten}
\usage{
gr.flatten(gr, gap = 0)
}
\arguments{
\item{gr}{TODO}

\item{gap}{TODO [0]}
}
\value{
TODO
}
\description{
Takes pile of GRanges and returns into a data.frame with nrow = length(gr) with each
representing the corresponding input range superimposed onto a single "flattened"
chromosome.
}

