% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.flatten}
\alias{gr.flatten}
\title{Lay ranges end-to-end onto a derivate "chromosome"}
\usage{
gr.flatten(gr, gap = 0)
}
\arguments{
\item{gr}{\code{GRanges} to flatten}

\item{gap}{Number of bases between ranges on the new chromosome \code{[0]}}
}
\value{
\code{data.frame} with start and end coordinates, and all of the original metadata
}
\description{
Takes pile of \code{GRanges} and returns into a \code{data.frame} with \code{nrow = length(gr)} with each
representing the corresponding input range superimposed onto a single "flattened"
chromosome, with ranges laid end-to-end
}

