% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gr.flatten}
\alias{gr.flatten}
\title{gr.flatten}
\usage{
gr.flatten(gr, gap = 0)
}
\arguments{
\item{gr}{\code{GRanges} to flatten}

\item{gap}{Default 0}
}
\description{
Takes pile of GRanges and returns into a data.frame with nrow = length(gr) with each
representing the corresponding input range superimposed onto a single "flattened"
chromosome.
}

