% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\docType{methods}
\name{\%&\%}
\alias{\%&\%}
\alias{\%&\%,}
\alias{GRanges-method}
\alias{\%&\%.GRangesList}
\alias{GRangesList-method}
\alias{\%&\%.CompressedGRangesList}
\alias{CompressedGRangesList-method}
\alias{\%^\%}
\alias{\%^\%,GRanges-method}
\alias{\%^^\%}
\alias{\%^^\%,GRanges-method}
\title{subset x on y ranges while ignoring strand (strand-agnostic)}
\usage{
\S4method{\%&\%}{GRanges}(x, y)

\S4method{\%&\%}{GRangesList}(x, y)

\S4method{\%&\%}{CompressedGRangesList}(x, y)

x \%^\% ...

x \%^^\% ...
}
\arguments{
\item{x}{See \link{gr.in}}

\item{...}{See \link{gr.in}}
}
\value{
subset of gr1 that overlaps gr2

subset of gr1 that overlaps gr2

subset of gr1 that overlaps gr2

logical vector of length gr1 which is TRUE at entry i only if gr1[i] intersects at least one interval in gr2 (strand agnostic)

logical vector of length gr1 which is TRUE at entry i only if gr1[i] intersects at least one interval in gr2
}
\description{
shortcut for x[gr.in(x,y)]

gr1 %&% gr2 returns the subsets of gr1 that overlaps gr2

shortcut for x[gr.in(x,y)]

gr1 %&% gr2 returns the subsets of gr1 that overlaps gr2

shortcut for x[gr.in(x,y)]

gr1 %&% gr2 returns the subsets of gr1 that overlaps gr2

Shortcut for gr.in (standard arguments)

gr1 %^% gr2

Shortcut for gr.in

gr1 %^^% gr2
}
\author{
Marcin Imielinski
}
