% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{\%&&\%}
\alias{\%&&\%}
\alias{\%&&\%,GRanges-method}
\title{Subset x on y ranges while respecting strand}
\usage{
x \%&&\% ...
}
\value{
subset of gr1 that overlaps gr2
}
\description{
shortcut for x[gr.in(x,y)]

gr1 %&&% gr2 returns the subsets of gr1 that overlaps gr2
}
\author{
Marcin Imielinski
}
