% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gr.in}
\alias{gr.in}
\title{gr.in}
\usage{
gr.in(query, subject, by = NULL, pintersect = FALSE, ...)
}
\arguments{
\item{query}{\code{GRanges} of query}

\item{subject}{\code{GRanges} of subject}

\item{by}{column name in query and subject that we additionally control for a match (passed on to gr.findoverlaps)}

\item{...}{params to send to \code{\link{gr.findoverlaps}}}
}
\description{
faster implementation of GRanges %over%  (uses gr.findoverlaps)
}
\details{
returns T / F vector if query range i is found in any range in subject

by = column name in query and subject that we additionally control for a match (passed on to gr.findoverlaps)
}

