% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.in}
\alias{gr.in}
\title{Faster version of \code{GRanges} \code{over}}
\usage{
gr.in(query, subject, ...)
}
\arguments{
\item{query}{Query \code{GRanges} pile}

\item{subject}{Subject \code{GRanges} pile}

\item{...}{Additional arguments to pass to \code{\link{gr.findoverlaps}}}
}
\value{
logical vector if query range i is found in any range in subject
}
\description{
Uses \code{\link{gr.findoverlaps}} for a faster \code{over}.
}
\details{
Can specify a \code{by} = column name in query and subject that we additionally control for a match (passed on to \code{\link{gr.findoverlaps}}).
}

