% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.in}
\alias{\%^\%}
\alias{\%^^\%}
\alias{gr.in}
\title{gr.in}
\usage{
gr.in(query, subject, by = NULL, pintersect = FALSE, ...)

x \%^\% ...

x \%^^\% ...
}
\arguments{
\item{query}{\code{GRanges} of query}

\item{subject}{\code{GRanges} of subject}

\item{by}{column name in query and subject that we additionally control for a match (passed on to gr.findoverlaps)}

\item{...}{params to send to \code{\link{gr.findoverlaps}}}
}
\value{
logical vector of length gr1 which is TRUE at entry i only if gr1[i] intersects at least one interval in gr2 (strand agnostic)

logical vector of length gr1 which is TRUE at entry i only if gr1[i] intersects at least one interval in gr2
}
\description{
faster implementation of GRanges %over%  (uses gr.findoverlaps)

returns T / F vector if query range i is found in any range in subject

by = column name in query and subject that we additionally control for a match (passed on to gr.findoverlaps)

Shortcut for gr.in (standard arguments)

gr1 %^% gr2

Shortcut for gr.in

gr1 %^^% gr2
}
\author{
Marcin Imielinski

Marcin Imielinski
}

