% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.in}
\alias{gr.in}
\title{Versatile implementation of \code{GenomicRanges::findOverlaps}}
\usage{
gr.in(query, subject, by = NULL, ...)
}
\arguments{
\item{query}{GRanges Set of GRanges to query. Refer to gr.findoverlaps() and GenomicRanges::findOverlaps()}

\item{subject}{GRanges Set of GRanges as 'subject' in query. Refer to gr.findoverlaps() and GenomicRanges::findOverlaps()}

\item{...}{Arguments to be passed to \code{\link{gr.findoverlaps}}}
}
\value{
boolean vector whereby TRUE is if query range i is found in any range in subject
}
\description{
Versatile implementation of \code{GenomicRanges::findOverlaps}

Returns boolean vector. TRUE if query range i is found in any range in subject.
}
