% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.match}
\alias{gr.match}
\title{Faster \code{GenomicRanges::match}}
\usage{
gr.match(query, subject, max.slice = Inf, verbose = FALSE, mc.cores = 1,
  ...)
}
\arguments{
\item{query}{Query \code{GRanges} pile}

\item{subject}{Subject \code{GRanges} pile}

\item{max.slice}{Maximum number of ranges to consider at once [Inf]}

\item{verbose}{Increase the verbosity during runtime}

\item{mc.cores}{Number of cores to use, if ranges exceed \code{max.slice}}

\item{...}{Additional arguments to be passed along to \code{gr.findoverlaps}}
}
\description{
Faster implementation of GRanges match (uses gr.findoverlaps)
returns indices of query in subject or NA if none found
... = additional args for findOverlaps (IRanges version)
}

