% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.match}
\alias{gr.match}
\title{gr.match}
\usage{
gr.match(query, subject, max.slice = Inf, verbose = FALSE, mc.cores = 1,
  ...)
}
\arguments{
\item{query}{\code{GRanges} object as query}

\item{subject}{\code{GRanges} object as subject}

\item{max.slice}{Default Inf. If query is bigger than this, chunk into smaller on different cores}

\item{verbose}{Default FALSE}

\item{mc.cores}{Default 1. Only works if exceeded max.slice}

\item{...}{arguments to be passed to \link{gr.findoverlaps}}
}
\value{
returns indices of query in subject or NA if none found
}
\description{
Faster implementation of GRanges match (uses gr.findoverlaps)
returns indices of query in subject or NA if none found
}

