% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.match}
\alias{gr.match}
\title{Faster \code{GenomicRanges::match}}
\usage{
gr.match(query, subject, max.slice = Inf, verbose = FALSE, mc.cores = 1,
  ...)
}
\arguments{
\item{query}{Query \code{GRanges} pile}

\item{subject}{Subject \code{GRanges} pile}

\item{max.slice}{Maximum number of ranges to consider at once [Inf]}

\item{verbose}{Increase the verbosity during runtime}

\item{mc.cores}{Number of cores to use, if ranges exceed \code{max.slice}}

\item{...}{Additional arguments to be passed along to \code{gr.findoverlaps}}
}
\value{
Vector of length = \code{length(query)} with subject indices of *first* subject in query, or NA if none found.
gThis behavior is different from \code{gr.findoverlaps}, which will
return *all* indicies of subject in query (in the case of one query overlaps with multiple subject)
... = additional args for findOverlaps (IRanges version)
}
\description{
Faster implementation of \code{GenomicRanges::match} (uses \code{gr.findoverlaps})
}

