% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.merge}
\alias{gr.merge}
\title{merge GRanges by using coordinates as primary key}
\usage{
gr.merge(query, subject, by = NULL, all = FALSE, all.query = all,
  all.subject = all, ignore.strand = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{query}{GRanges Set of GRanges to query. Refer to gr.findoverlaps() and GenomicRanges::findOverlaps()}

\item{subject}{GRanges Set of GRanges as 'subject' in query. Refer to  gr.findoverlaps() and GenomicRanges::findOverlaps()}

\item{by}{vector Additional metadata fields to join on}

\item{all}{boolean Flag whether to include left and right joins}

\item{all.query}{boolean Flag whether to do a left join (default = all)}

\item{all.subject}{boolean Flag whether to do a right join (default = all)}

\item{ignore.strand}{boolean Strand information ignored (default = TRUE)}

\item{verbose}{boolean}
}
\value{
GRanges merged on 'by' vector
}
\description{
Uses gr.findoverlaps() to enable internal and external joins of GRanges using
syntax similar to "merge" where merging is done using coordinates +/- "by" fields

Uses gr.findoverlaps() / GRanges::findOverlaps for heavy lifting, but returns outputs with
metadata populated as well as query and subject ids.  For external joins,
overlaps x with gaps(y) and gaps(x) with y.
}
