% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.mid}
\alias{gr.mid}
\title{Get the midpoint of range}
\usage{
gr.mid(x)
}
\arguments{
\item{x}{\code{GRanges} object to operate on}

\item{pad}{Amount to pad the output width to (default 0, so width of 1)}
}
\value{
\code{GRanges} of the midpoint, calculated from \code{floor(width(x)/2)}
}
\description{
Get the midpoint of range
}
\examples{
\dontrun{gr.mid(GRanges(1, IRanges(1000,2000), seqinfo=Seqinfo("1", 2000)))}
}

