% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.nochr}
\alias{gr.nochr}
\title{Remove chr prefix from GRanges seqlevels}
\usage{
gr.nochr(gr)
}
\arguments{
\item{gr}{GRanges with chr seqlevel prefixes}
}
\value{
GRanges without chr seqlevel prefixes
}
\description{
Remove chr prefix from GRanges seqlevels
}

