% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.noval}
\alias{gr.noval}
\title{get rid of mcols on GRanges/GRangesLists}
\usage{
gr.noval(gr, keep.col = NULL, drop.col = NULL)
}
\value{
GRanges or GRangesList
}
\description{
remove all metadata from GRanges or GRangesList
}
\author{
Kevin Hadi
}
