% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{\%+\%}
\alias{\%+\%}
\alias{\%+\%,GRanges-method}
\title{Nudge GRanges right}
\usage{
gr \%+\% ...
}
\value{
shifted granges
}
\description{
Operator to shift GRanges right "sh" bases
}
\author{
Marcin Imielinski
}
