% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{\%+\%}
\alias{\%+\%}
\alias{\%-\%}
\title{Nudge GRanges right}
\usage{
gr \%+\% ...

gr \%-\% ...
}
\value{
shifted granges

shifted granges
}
\description{
Operator to shift GRanges right "sh" bases

Operator to shift GRanges left "sh" bases

df %!% c('string.*to.*match', 'another.string.to.match')
}
\author{
Marcin Imielinski

Marcin Imielinski
}

