% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{\%-\%}
\alias{\%-\%}
\title{Shift GRanges left}
\usage{
gr \%-\% x
}
\value{
shifted GRanges
}
\description{
Operator to shift GRanges left "sh" bases

df %!% c('string.*to.*match', 'another.string.to.match')
}
\author{
Marcin Imielinski
}
