% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.pairflip}
\alias{gr.pairflip}
\title{Create pairs of ranges and their strand-inverse}
\usage{
gr.pairflip(gr)
}
\arguments{
\item{gr}{\code{GRanges}}
}
\value{
\code{GRangesList} with each element of length 2
}
\description{
From a \code{GRanges} returns a \code{GRangesList} with each item consisting
of the original \code{GRanges} and its strand flip
}
