% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.reduce}
\alias{gr.reduce}
\title{Minimal overlaps for GRanges/GRangesList}
\usage{
gr.reduce(...)
}
\arguments{
\item{...}{\code{GRanges} or \code{GRangesList}}
}
\value{
GRanges
}
\description{
Takes any number of GRanges or GRangesList and reduces them to the minimal
set of overlapping windows, ignoring strand.
}

