% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{\%_\%}
\alias{\%_\%}
\alias{\%_\%,GRanges-method}
\title{\code{BiocGenerics::setdiff} shortcut (strand agnostic)}
\usage{
x \%_\% ...
}
\arguments{
\item{x}{\code{GRanges} object to to}

\item{...}{A \code{GRanges} or a character to be parsed into a \code{GRanges}}
}
\value{
\code{GRanges} representing setdiff of input interval
}
\description{
Shortcut for \code{BiocGenerics::setdiff}

gr1 <- GRanges(1, IRanges(10,20), strand="+")
gr2 <- GRanges(1, IRanges(15,25), strand="-")
gr3 <- "1:1-15"
gr1 %_% gr2
gr1 %_% gr3
}
\author{
Marcin Imielinski
}

