% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{\%_\%}
\alias{\%_\%}
\alias{\%__\%}
\title{setdiff shortcut (strand agnostic)}
\usage{
x \%_\% ...

x \%__\% ...
}
\value{
granges representing setdiff of input interval

granges representing setdiff of input interval
}
\description{
Shortcut for union

gr1 %_% gr2

Shortcut for setdiff

gr1 %__% gr2
}
\author{
Marcin Imielinski

Marcin Imielinski
}

