% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{\%_\%}
\alias{\%_\%}
\alias{\%_\%,GRanges-method}
\alias{gr.setdiff}
\title{\code{BiocGenerics::setdiff} shortcut (strand agnostic)}
\usage{
x \%_\% ...

gr.setdiff(query, subject, ignore.strand = TRUE, by = NULL, ...)
}
\arguments{
\item{x}{\code{GRanges} object to to}

\item{...}{A \code{GRanges} or a character to be parsed into a \code{GRanges}}

\item{query}{\code{GRanges} object as query}

\item{subject}{\code{GRanges} object as subject}

\item{ignore.strand}{boolean Flag to ignore strands. Refer to 'gr.findoverlaps()'. (default = TRUE)}

\item{by}{vector Meta-data column to consider when performing overlaps. Refer to 'gr.findoverlaps()' documentation (default = NULL)}

\item{...}{arguments to be passed to \link{gr.findoverlaps}}
}
\value{
\code{GRanges} representing setdiff of input interval

Returns indices of query in subject. If none found, NA
}
\description{
Shortcut for \code{BiocGenerics::setdiff}

gr1 <- GRanges(1, IRanges(10,20), strand="+")
gr2 <- GRanges(1, IRanges(15,25), strand="-")
gr3 <- "1:1-15"
gr1 %_% gr2
gr1 %_% gr3

More robust and faster implementation of GenomicRanges::setdiff()

Robust to common edge cases of setdiff(gr1, gr2)  where gr2 ranges are contained inside gr1's (yieldings
setdiffs yield two output ranges for some of the input gr1 intervals.
}
\author{
Marcin Imielinski
}
