% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.setdiff}
\alias{gr.setdiff}
\title{More robust and faster implementation of \code{GenomicRanges::setdiff}}
\usage{
gr.setdiff(query, subject, ignore.strand = TRUE, by = NULL, ...)
}
\arguments{
\item{query}{\code{GRanges} object as query}

\item{subject}{\code{GRanges} object as subject}

\item{ignore.strand}{Don't consider strand information [TRUE]}

\item{by}{Additional meta.data to consider when finding set differences [NULL]}

\item{...}{Additional arguments to be passed to \code{gr.findoverlaps}}
}
\value{
returns indices of query in subject or NA if none found
}
\description{
Robust to common edge cases of setdiff(gr1, gr2)  where gr2 ranges are contained inside gr1's (yieldings
setdiffs yield two output ranges for some of the input gr1 intervals.
}

