% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.simplify}
\alias{gr.simplify}
\title{Simplify granges by collapsing all non-overlapping adjacent ranges that share a given "field" value
(adjacent == adjacent in the input GRanges object)}
\usage{
gr.simplify(gr, field = NULL, val = NULL, include.val = T, split = F,
  pad = 1)
}
\arguments{
\item{gr}{takes in gr or grl}

\item{field}{character scalar, corresponding to value field of gr}

\item{include.val}{scalar logical, will include in out gr values field of first matching record in input gr}
}
\value{
Simplified GRanges with "field" populated with uniquely contiguous values
}
\description{
Simplify granges by collapsing all non-overlapping adjacent ranges that share a given "field" value
(adjacent == adjacent in the input GRanges object)
}

