% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.simplify}
\alias{gr.simplify}
\title{Simplify \code{GRanges} by collapsing all non-overlapping adjacent ranges that share a given "field" value}
\usage{
gr.simplify(gr, field = NULL, val = NULL, include.val = TRUE,
  split = FALSE, pad = 1)
}
\arguments{
\item{gr}{takes in gr or grl}

\item{field}{character scalar, corresponding to value field of gr}

\item{val}{TODO}

\item{include.val}{scalar logical, will include in out gr values field of first matching record in input gr}

\item{split}{TODO}

\item{pad}{TODO [1]}
}
\value{
Simplified GRanges with "field" populated with uniquely contiguous values
}
\description{
Note that adjacent == adjacent in the input GRanges object
}

