% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gr.string}
\alias{gr.string}
\title{gr.string}
\usage{
gr.string(gr, add.chr = F, mb = F, round = 3, other.cols = c())
}
\description{
return ucsc style interval string corresponding to gr pile (ie chr:start-end)
}
\details{
if mb will return as MB and round to "round"
}

