% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.string}
\alias{gr.string}
\title{Return UCSC style interval string corresponding to \code{GRanges} pile (ie chr:start-end)}
\usage{
gr.string(gr, add.chr = FALSE, mb = FALSE, round = 3, other.cols = c())
}
\arguments{
\item{gr}{\code{GRanges} pile to get intervals from}

\item{add.chr}{Prepend seqnames with "chr" \code{[FALSE]}}

\item{mb}{Round to the nearest megabase \code{[FALSE]}}

\item{round}{If \code{mb} supplied, how many digits to round to. \code{[3]}}

\item{other.cols}{Names of additional \code{mcols} fields to add to the string (seperated by ";")}
}
\description{
Return UCSC style interval string corresponding to \code{GRanges} pile (ie chr:start-end)
}
\examples{
gr.string(gr.genes, other.cols = c("name", "name2"))
}

