% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.string}
\alias{gr.string}
\title{gr.string}
\usage{
gr.string(gr, add.chr = FALSE, mb = FALSE, round = 3, other.cols = c(),
  pretty = FALSE)
}
\description{
return ucsc style interval string corresponding to gr pile (ie chr:start-end)
}
\details{
if mb will return as MB and round to "round"
}

