% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.tile.map}
\alias{gr.tile.map}
\title{gr.tile.map}
\usage{
gr.tile.map(query, subject, verbose = FALSE)
}
\arguments{
\item{query}{Query \code{GRanges} pile, perhaps created from some tile (e.g. \code{gr.tile}), and assumed to have no gaps}

\item{subject}{Subject \code{GRanges} pile, perhaps created from some tile (e.g. \code{gr.tile}), and assumed to have no gaps}

\item{verbose}{Increase the verbosity of the output (default = FALSE)}
}
\value{
\code{list} of length = \code{length(query)}, where each element \code{i} is a vector of indicies in \code{subject} that overlaps element \code{i} of \code{query}
}
\description{
Given two tilings of the genome (e.g. at different resolution)
query and subject outputs a length(query) list whose items are integer vectors of indices in subject
overlapping that overlap that query (strand non-specific)
}
\note{
Assumes that input query and subject have no gaps (including at end) or overlaps, i.e. ignores end()
coordinates and only uses "starts"
}
\author{
Marcin Imielinski
}
