% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr.tile.map}
\alias{gr.tile.map}
\title{gr.tile.map}
\usage{
gr.tile.map(query, subject, mc.cores = 1, verbose = FALSE)
}
\arguments{
\item{query}{Query \code{GRanges} pile}

\item{subject}{Subject \code{GRanges} pile}

\item{mc.cores}{Number of cores to us [1]}

\item{verbose}{Increase the verbosity of the output}
}
\value{
TODO
}
\description{
Given two tilings of the genome (e.g. at different resolution)
query and subject outputs a length(query) list whose items are integer vectors of indices in subject
overlapping that overlap that query (strand non-specific)
}
\note{
Assumes that input query and subject have no gaps (including at end) or overlaps, i.e. ignores end()
coordinates and only uses "starts"
}

