% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gr.tile}
\alias{gr.tile}
\title{gr.tile}
\usage{
gr.tile(gr, w = 1000)
}
\arguments{
\item{gr}{GRanges to tile, also can be seqlengths or seqinfo}

\item{w}{Width of the binds. Default 1e3}
}
\description{
tiles interval (or whole genome) with segments of <= specified width.  Returns strandless gr
"tiles".
}
\details{
input can be seqinfo object (in which case whole genome will be tiled);
if inputted grs overlap, will first reduce then tile.
}

