% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\docType{methods}
\name{\%N\%}
\alias{\%$\%}
\alias{\%$\%,GRanges-method}
\alias{\%N\%}
\alias{\%N\%,GRanges-method}
\title{gr.val shortcut to get total numbers of intervals in gr2 overlapping with each interval in  gr1, ignoring strand}
\usage{
x \%N\% ...

x \%$\% ...
}
\arguments{
\item{x}{See \link{gr.val}}

\item{...}{See \link{gr.val}}

\item{x}{\code{GRanges} object}
}
\value{
bases overlap of gr1 with gr2

gr1 with extra meta data fields populated from gr2
}
\description{
gr1 %N% gr2

Shortcut for gr.val (using val = names(values(y)))

gr1 %$% gr2
}
\author{
Marcin Imielinski

Marcin Imielinski
}

