% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr2dt}
\alias{gr2dt}
\title{Converts \code{GRanges} to \code{data.table}}
\usage{
gr2dt(gr)
}
\arguments{
\item{gr}{\code{GRanges} pile to convert to \code{data.table}}
}
\value{
\code{data.table} with seqnames, start, end, width, strand and all of the meta data. Width is end-inclusive (e.g. [6,7] width = 2)
}
\description{
Converts \code{GRanges} to \code{data.table}
}
\examples{
gr2dt(example_genes)
}

