% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{gr2grl}
\alias{gr2grl}
\title{gr2grl
Quick way to make grl from list of indices into a GRanges gr}
\usage{
gr2grl(gr, ix)
}
\description{
gr2grl
Quick way to make grl from list of indices into a GRanges gr
}

