% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{grbind}
\alias{grbind}
\title{Concatenate \code{GRanges}, robust to different \code{mcols}}
\usage{
grbind(x, ...)
}
\arguments{
\item{x}{First \code{GRanges}}

\item{...}{additional \code{GRanges}}
}
\value{
Concatenated \code{GRanges}
grbind(example_genes, example_dnase)
}
\description{
Concatenates \code{GRanges} objects, taking the union of their features if they have non-overlapping features
}
\note{
Does not fill in the \code{Seqinfo} for the output \code{GRanges}
}

