% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{\%**\%}
\alias{\%**\%}
\title{gr.findoverlaps (respects strand)}
\usage{
x \%**\% ...
}
\value{
new granges containing every pairwise intersection of ranges in gr1 and gr2 with a join of the corresponding metadata
}
\description{
Shortcut for gr.findoverlaps

gr1 %**% gr2
}
\author{
Marcin Imielinski
}

