% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\docType{methods}
\name{\%*\%}
\alias{\%*\%}
\alias{\%*\%,GRanges-method}
\title{Metadata join with coordinates as keys (wrapper to \code{\link{gr.findoverlaps}})}
\usage{
\S4method{\%*\%}{GRanges}(x, y)
}
\arguments{
\item{x}{\code{GRanges}}

\item{y}{\code{GRanges}}
}
\value{
\code{GRanges} containing every pairwise intersection of ranges in \code{x} and \code{y} with a join of the corresponding  metadata
}
\description{
Shortcut for gr.findoverlaps with \code{qcol} and \code{scol} filled in with all the query and subject metadata names.
This function is useful for piping \code{GRanges} operations together. Another way to think of %*% is as a
join of the metadata, with genomic coordinates as the keys. \cr
Example usage: \cr
x %*% y
}
\examples{
example_genes \%*\% example_dnase
}
\author{
Marcin Imielinski
}

