% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gUtils.R
\docType{methods}
\name{\%*\%}
\alias{\%*\%}
\title{gr.findoverlaps}
\usage{
\S4method{\%*\%}{GRanges,ANY}(x, y)
}
\value{
new granges containing every pairwise intersection of ranges in gr1 and gr2 with a join of the corresponding  metadata
}
\description{
Shortcut for gr.findoverlaps (standard arguments)

gr1 %*% gr2
}
\author{
Marcin Imielinski
}

