% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{grl.allin}
\alias{grl.allin}
\alias{grl.in}
\title{grl.allin}
\usage{
grl.in(grl, windows, some = FALSE, only = FALSE)
}
\arguments{
\item{grl}{\link{GRangesList} object to query}

\item{windows}{\link{GRanges} windows to check against}

\item{some}{[Default FALSE]}

\item{only}{[Default FALSE]}
}
\description{
Like %in% for grl but now will return a logical vector that is true at position if i
only if the ranges in grl[i] intersect <<all>>, <<some>>, <<only>>  windows in the subject
}
\details{
eg can use to identify read pairs whose ends are contained inside two genes)
}

