% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{grl.end}
\alias{grl.end}
\title{grl.end}
\usage{
grl.end(grl, width = 1, force = FALSE, ignore.strand = TRUE,
  clip = TRUE)
}
\arguments{
\item{grl}{something here}

\item{width}{someting here}

\item{force}{boolean someting here}

\item{ignore.strand}{boolean someting here}

\item{clip}{boolean someting here}
}
\description{
details here
}
