% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{grl.eval}
\alias{grl.eval}
\title{Evaluate and aggregate expression on GRanges column in GRangesList}
\usage{
grl.eval(grl, expr, condition = NULL)
}
\arguments{
\item{grl}{GRangesList to evaluate over}

\item{expr}{Any syntactically valid R expression, on columns of GRanges or GRangesList}

\item{condition}{Optional: any syntactically valid R expression, on columns of GRanges or GRangesList, on which to subset before evaluating main 'expr' (default = NULL)}
}
\value{
GRangesList evaluated by R expressions
}
\description{
Evaluate expression 'expr' on indivdual GRanges inside GRangesList.
Expression should result in a single i.e. scalar value per GRangesList item.
}
