% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{grl.eval}
\alias{grl.eval}
\title{evaluate and aggregate expression on GRanges column in GRangesList}
\usage{
grl.eval(grl, expr, condition = NULL)
}
\arguments{
\item{grl}{GRangesList to eval over}

\item{expr}{expression on columns of granges or granges list}

\item{condition}{optional expression (with logical or integer output) on columns of GRanges on which to subset prior to evaluating main expr}
}
\description{
Evaluate expression expr on indivdual granges inside grangeslist.
Expression should result in a single i.e. scalar value per grangeslist item.
}
