% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{grl.filter}
\alias{grl.filter}
\title{Filters GRangesList to only include ranges in the specified window}
\usage{
grl.filter(grl, windows)
}
\description{
(this is different from %in% which does not remove non matching ranges from the grls)
}
\details{
does not return list in necessarily same order
}

