% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{grl.pivot}
\alias{grl.pivot}
\title{Pivot a \code{GRangesList}, inverting "x" and "y"}
\usage{
grl.pivot(x)
}
\arguments{
\item{x}{\code{GRangesList} object to pivot}
}
\value{
GRangesList with inverted 'x' and 'y'
}
\description{
"Pivots" GRangesList object "x" by returning a new GRangesList "y" whose
kth item is GRanges object of ranges x[[i]][k] for all i in 1:length(x)

e.g. If the length of a GRangesList `grl` is 50, `length(grl)=50  and length of each \code{GRanges} element inside is 2,
then the function \code{grl.pivot} will produce a length 3 \code{GRangesList} with 50 elements per \code{GRanges}

Note: Assumes all GRanges in "x" are of equal length
}
\examples{
grl.pivot(grl.hiC)
}
