% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{grl.pivot}
\alias{grl.pivot}
\title{Pivot a \code{GRangesList}, inverting "x" and "y"}
\usage{
grl.pivot(x)
}
\arguments{
\item{x}{\code{GRangesList} object to pivot}
}
\description{
"Pivots" grl object "x" by returning a new grl "y" whose
kth item is gr object of ranges x[[i]][k] for all i in 1:length(x)
e.g. If \code{length(grl)} is 50 and length of each \code{GRanges} element inside is 2, then \code{grl.pivot}
will produce a length 3 \code{GRangesList} with 50 elements per \code{GRanges}
}
\details{
Assumes all grs in "x" are of equal length
}
\examples{
grl.pivot(grl.hiC)
}

