% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{grl.reduce}
\alias{grl.reduce}
\title{grl.reduce}
\usage{
grl.reduce(grl, pad = 0, clip = FALSE)
}
\arguments{
\item{grl}{\code{GRangesList}}

\item{pad}{padding to add to ranges inside grl before reduing}
}
\value{
\code{GRangesList} with reduced intervals
}
\description{
Quickly ranges inside grl +/- pad
Can use with split / unlist
}
\author{
Marcin Imielinski
}
