% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{grl.start}
\alias{grl.start}
\title{Get the left ends of a \code{GRangesList}}
\usage{
grl.start(grl, width = 1, force = FALSE, ignore.strand = TRUE, clip = TRUE)
}
\arguments{
\item{grl}{\code{GRangesList} object containing GRanges to operate on}

\item{width}{integer Specify subranges of greater width including the start of the range. (default = 1)}

\item{force}{boolean Allows returned \code{GRangesList} to have ranges outside of its \code{Seqinfo} bounds. (default = FALSE)}

\item{ignore.strand}{boolean If set to \code{FALSE}, will extend '-' strands from the other direction (default = TRUE)}

\item{clip}{boolean Trims returned \code{GRangesList} so that it does not extend beyond bounds of the input \code{GRanges} (default = TRUE)}
}
\value{
\code{GRangesList} object of width 1 ranges representing start of each genomic range in the input.

\code{GRangesList} object of width 1+ ranges representing start of each genomic range in the input.
}
\description{
Function wrapping around gr.start to work
on GRangesLists
}
