% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{grl.string}
\alias{grl.string}
\title{grl.string}
\usage{
grl.string(grl, sep = ",", ...)
}
\arguments{
\item{grl}{GenomicRangesList to get strings from}

\item{sep}{[Default ","]}

\item{...}{dummy}
}
\description{
return ucsc style interval string corresponding to each gr in each grl, one line per grl item, with gr's in
grl separated by sep
}
\details{
if mb will return as MB and round to "round"
}

