% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{grl.string}
\alias{grl.string}
\title{Create string representation of \code{GRangesList}}
\usage{
grl.string(grl, mb = FALSE, sep = ",", ...)
}
\arguments{
\item{grl}{\code{GRangesList} to convert to string vector}

\item{mb}{Will return as MB and round to "round" [FALSE]}

\item{sep}{Character to separate single \code{GRanges} ranges [,]}

\item{...}{Additional arguments to be passed to \code{gr.string}}
}
\value{
Character vector where each element is a \code{GRanges} pile corresponding to a single \code{GRangesList} element
}
\description{
Return ucsc style interval string corresponding to each \code{GRanges} in the \code{GRangesList}.
One line per per \code{GRangesList} item. \code{GRanges} elements themselves are separated by \code{sep}
}
\examples{
grl.string(grl.hiC, mb=TRUE)
}

