% Generated by roxygen2 (4.0.2): do not edit by hand
\name{grl.string}
\alias{grl.string}
\title{grl.string}
\usage{
grl.string(grl, sep = ",", ...)
}
\description{
return ucsc style interval string corresponding to each gr in each grl, one line per grl item, with gr's in
grl separated by sep
}
\details{
if mb will return as MB and round to "round"
}

