% Generated by roxygen2 (4.0.2): do not edit by hand
\name{grl.unlist}
\alias{grdt}
\alias{grl.unlist}
\title{grl.unlist}
\usage{
grl.unlist(grl)

grdt(x)
}
\description{
Does a "nice" unlist of a grl object adding a field "grl.ix" denoting which element of the grl
each gr corresponds to

Converts gr to data frame
}
\details{
and a field grl.iix which saves the (local) index that that gr was in its corresponding grl item

and a field grl.iix which saves the (local) index that that gr was in its corresponding grl item
}

