% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{grl.unlist}
\alias{grl.unlist}
\title{grl.unlist}
\usage{
grl.unlist(grl)
}
\description{
Does a "nice" unlist of a grl object adding a field "grl.ix" denoting which element of the grl
each gr corresponds to
}
\details{
and a field grl.iix which saves the (local) index that that gr was in its corresponding grl item
}

