% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hg_seqlengths}
\alias{hg_seqlengths}
\title{Output standard human genome seqlengths}
\usage{
hg_seqlengths(hg19 = TRUE, chr = FALSE, include.junk = FALSE)
}
\arguments{
\item{hg19}{Flag for dealing with hg19 (vs hg18). Default TRUE}

\item{chr}{Flag for whether to keep "chr". Default FALSE}

\item{include.junk}{Flag for whether to not trim to only 1-22, X, Y, M. Default FALSE}
}
\value{
Seqlengths
}
\description{
Outputs a standard seqlengths for human genome +/- "chr"
}

