% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{hg_seqlengths}
\alias{hg_seqlengths}
\title{Output standard human genome seqlengths}
\usage{
hg_seqlengths(hg, chr = FALSE, include.junk = FALSE)
}
\arguments{
\item{hg}{BSgenome object, perhaps from \code{library(BSgenome.Hsapiens.UCSC.hg19)}}

\item{chr}{Flag for whether to keep "chr". Default FALSE}

\item{include.junk}{Flag for whether to not trim to only 1-22, X, Y, M. Default FALSE}
}
\value{
Seqlengths
}
\description{
Outputs a standard seqlengths for human genome +/- "chr"
}

