% Generated by roxygen2 (4.0.2): do not edit by hand
\name{import.ucsc}
\alias{import.ucsc}
\title{import.ucsc}
\usage{
import.ucsc(con, selection = NULL, text, chrsub = T, verbose = FALSE,
  as = NULL, ...)
}
\description{
wrapper around rtracklayer import that
(1) handles "as" formats
(2) has additional flag chrsub to sub in 'chr' in selection, and then sub it out of the output
}

