% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mafcount}
\alias{mafcount}
\title{mafcount}
\usage{
mafcount(tum.bam, norm.bam = NULL, maf, chunk.size = 100, verbose = T,
  mc.cores = 1, ...)
}
\description{
Returns base counts for reference and alternative allele for an input tum and norm bam and maf data frame or GRAnges specifying substitutions
}
\details{
maf is a single width GRanges describing variants and field 'ref' (or 'Reference_Allele'), 'alt' (or 'Tum_Seq_Allele1') specifying reference and alt allele.
maf is assumed to have width 1 and strand is ignored.
}

