% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{parse.gr}
\alias{parse.gr}
\title{Parse a string into \code{GRanges}}
\usage{
parse.gr(...)
}
\arguments{
\item{...}{arguments to \link{parse.grl} i.e. character strings in UCSC style chr:start-end[+-]}
}
\value{
\code{GRanges} objects parsed from input
}
\description{
Parses character vector with IGV / UCSC style strings of format gr1;gr2;gr3,
where each \code{GRanges} is of format \code{chr:start-end[+/-]}
}

