% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{parse.gr}
\alias{parse.gr}
\title{parse.gr}
\usage{
parse.gr(...)
}
\arguments{
\item{...}{arguments to parse.grl i.e. character strings in UCSC style chr:start-end[+-]}
}
\description{
Quick function to parse gr from character vector IGV / UCSC style strings of format gr1;gr2;gr3 wohere each gr is of format chr:start-end[+/-]
}
\author{
Marcin Imielinski
}
