% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{parse.grl}
\alias{parse.grl}
\title{parse.grl}
\usage{
parse.grl(x, seqlengths = hg_seqlengths())
}
\description{
quick function to parse \code{GRangesList} from character vector IGV / UCSC style strings of format gr1;gr2;gr3 where each gr is of format chr:start-end[+/-]
}

