% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parse.grl}
\alias{gstring}
\alias{parse.grl}
\title{parse.grl}
\usage{
parse.grl(x, seqlengths = hg_seqlengths())

gstring(...)
}
\description{
quick function to parse \code{GRangesList} from character vector IGV / UCSC style strings of format gr1;gr2;gr3 where each gr is of format chr:start-end[+/-]

quick function to parse gr from character vector IGV / UCSC style strings of format gr1;gr2;gr3 where each gr is of format chr:start-end[+/-]
}

