% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{ra.overlaps}
\alias{ra.overlaps}
\title{ra.overlaps}
\usage{
ra.overlaps(ra1, ra2, pad = 0, arr.ind = T, ignore.strand = FALSE, ...)
}
\description{
Determines overlaps between two piles of rearrangement junctions ra1 and ra2 (each GRangesLists of signed locus pairs)
against each other, returning a sparseMatrix that is T at entry ij if junction i overlaps junction j.
}
\details{
if argument pad = 0 (default) then only perfect overlap will validate, otherwise if pad>0 is given, then
padded overlap is allowed

strand matters, though we test overlap of both ra1[i] vs ra2[j] and gr.flip(ra2[j])
}

