% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ra.overlaps}
\alias{ra.overlaps}
\alias{ra.setdiff}
\title{ra.overlaps}
\usage{
ra.overlaps(ra1, ra2, pad = 0, arr.ind = T, ignore.strand = FALSE, ...)

ra.setdiff(ra1, ra2, ...)
}
\description{
Determines overlaps between two piles of rearrangement junctions ra1 and ra2 (each GRangesLists of signed locus pairs)
against each other, returning a sparseMatrix that is T at entry ij if junction i overlaps junction j.

identifies events that are in ra1 that do not exist in ra2.
Aside from ra1 and ra2, all arguments are sent to ra.overlaps
}
\details{
if argument pad = 0 (default) then only perfect overlap will validate, otherwise if pad>0 is given, then
padded overlap is allowed

strand matters, though we test overlap of both ra1[i] vs ra2[j] and gr.flip(ra2[j])
}

