% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ra_breaks}
\alias{ra_breaks}
\title{ra_breaks}
\usage{
ra_breaks(rafile, keep.features = TRUE, seqlengths = hg_seqlengths(),
  chr.convert = TRUE, snowman = FALSE, breakpointer = FALSE,
  seqlevels = NULL, get.loose = FALSE)
}
\arguments{
\item{rafile}{dRanger or Snowman file to read in}

\item{keep.features}{[Default TRUE] keep metadata features from VCF etc as metadata}

\item{seqlengths}{Seqlengths to be used for this. Default \link{hg_seqlengths}}

\item{chr.convert}{dummy}

\item{snowman}{is this a snowman file. Default FALSE}

\item{breakpointer}{is this a breakpointer file. Default FALSE}

\item{seqlevels}{dummy}

\item{get.loose}{[Default FALSE] if TRUE will return a list with fields $junctions and $loose.ends}
}
\description{
takes in either file or data frame from dranger or snowman or path to BND / SV type vcf file
and returns junctions in VCF format.
}
\details{
The default output is GRangesList each with a length two GRanges whose strands point AWAY from the break.  If get.loose = TRUE (only relevant for VCF)
}

