% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_hg}
\alias{read_hg}
\title{Wrapper around BSgenome call}
\usage{
read_hg(hg19 = T, fft = F)
}
\arguments{
\item{hg19}{Logical whether to return hg18 or hg19 BSgenome. Default TRUE}

\item{fft}{Logical whether to return an ffTrack. Default FALSE}
}
\value{
BSGenome or ffTrack of the genome
}
\description{
Retreives either the BSgenome hg18 or hg19 genome.
Can also return the hg19 ffTrack (requires that the file exists)
}

