% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rle.query}
\alias{rle.query}
\title{rle.query}
\usage{
rle.query(subject.rle, query.gr, verbose = FALSE, mc.cores = 1,
  chunksize = 1e+09)
}
\arguments{
\item{subject.rle}{Subject}

\item{query.gr}{Query}

\item{verbose}{Default FALSE}

\item{mc.cores}{number of cores}

\item{chunksize}{Default 1e9}
}
\value{
Rle representing the (concatenated) vector of data (reversing order in case of negative strand input)
}
\description{
Queries an \code{\link{RleList}} representing genomic data (ie a list whose names represent
seqnames ie chromosomes, and lengths represent seqlengths)
via \code{GRanges} object
}
\note{
Throws warning if seqlengths(gr) do not correspond to the lengths of the \code{RleList} components
}

