% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{rrbind}
\alias{rrbind}
\title{Improved \code{rbind} for intersecting/union columns of \code{data.frames} or \code{data.tables}}
\usage{
rrbind(..., union = TRUE, as.data.table = FALSE)
}
\arguments{
\item{...}{Any number of \code{data.frame} or \code{data.table} objects}

\item{union}{Take union of columns (and put NA's for columns of df1 not in df2 and vice versa). (default == TRUE)}

\item{as.data.table}{Return the binded data as a \code{data.table}. (default == FALSE)}
}
\value{
\code{data.frame} or \code{data.table} of the \code{rbind} operation
}
\description{
Improved \code{rbind} for intersecting/union columns of \code{data.frames} or \code{data.tables}

Like \code{rbind}, but takes the intersecting columns of the data.
}
\author{
Marcin Imielinski
}
