% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{seg2gr}
\alias{seg2gr}
\title{Convert GRange like data.frames into GRanges}
\usage{
seg2gr(segs, seqlengths = NULL, seqinfo = Seqinfo())
}
\arguments{
\item{segs}{data frame of segments with fields denoting chromosome, start, end, and other metadata (see standardized segs for seg data frame input formats)}

\item{seqlengths}{seqlengths of output GRanges object}

\item{seqinfo}{seqinfo of output GRanges object}
}
\description{
Take data frame of ranges "segs" and converts into granges object porting over additional value columns
"segs" data frame can obey any number of conventions to specify chrom, start, and end of ranges
(eg $pos1, $pos2, $Start_position, $End_position) --> see "standardize_segs" for more info
}

