% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{si2gr}
\alias{si2gr}
\title{Shortcut GRanges from Seqinfo}
\usage{
si2gr(...)
}
\arguments{
\item{si}{Seqinfo object}

\item{strip.empty}{Don't know. Default FALSE}
}
\description{
Creates a genomic ranges from seqinfo object
ie a pile of ranges spanning the genome
}
\examples{
\dontrun{si <- Seqinfo(names(hg_seqlength(), hg_seqlengths()))
si2gr(si)}
}

