% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{standardize_segs}
\alias{standardize_segs}
\title{standardize_segs}
\usage{
standardize_segs(seg, chr = F)
}
\description{
(data frame seg function)
}
\details{
Takes and returns segs data frame standardized to a single format (ie $chr, $pos1, $pos2)
if chr = T will ensure "chr" prefix is added to chromossome(if does not exist)
}

