% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{standardize_segs}
\alias{standardize_segs}
\title{Takes and returns segs data frame standardized to a single format}
\usage{
standardize_segs(seg, chr = FALSE)
}
\arguments{
\item{seg}{data.frame or data.table of segments with fields denoting chromosome, start, end, and other metadata.}

\item{chr}{boolean Flag to force add chromosomes (default = FALSE)}
}
\value{
data.frame or data.table with standardized segments
}
\description{
Takes and returns segs data frame standardized to a single format (i.e. $chr, $pos1, $pos2)

If chr = TRUE will ensure "chr" prefix is added to chromossome(if does not exist)

"segs" data.frame can obey any number of conventions to specify chrom, start, and end of ranges
(e.g. $pos1, $pos2, $Start_position, $End_position) 

Conventions:
\itemize{
\item \code{ID} - 'id', 'patient', 'sample'
\item \code{chr} - 'seqnames', 'chrom', 'chromosome', 'rname', 'space', 'contig'
\item \code{pos1} - 'start', 'loc.start', 'start.bp', 'start_position', 'begin', 'pos', 'pos1', 'left', 's1'
\item \code{pos2} - 'end', 'loc.end', 'stop', 'end.bp', 'end_posiiton', 'pos2', 'right', 'e1'
\item s\code{strand} - 'strand', 'str' 
}
}
