% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gUtils.R
\name{streduce}
\alias{streduce}
\title{Reduce \code{GRanges} and \code{GRangesList} to miminal footprint}
\usage{
streduce(gr, pad = 0, sort = TRUE)
}
\arguments{
\item{gr}{\code{GRanges} or \code{GRangesList}}

\item{pad}{Expand the input data before reducing. \code{[0]}}

\item{sort}{Flag to sort the output. \code{[TRUE]}}
}
\value{
\code{GRanges} object with no strand information, representing a minimal footprint
}
\description{
Shortcut for \code{reduce(sort(gr.stripstrand(unlist(x))))}
}
\examples{
streduce(grl.hiC, pad=10)
streduce(example_genes, pad=1000)
}

