% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{streduce}
\alias{streduce}
\title{Shortcut for \code{reduce(sort(gr.stripstrand(unlist(x))))}}
\usage{
streduce(gr, pad = 0, sort = TRUE)
}
\arguments{
\item{gr}{takes in gr or grl}

\item{pad}{asdf. Default 0}

\item{sort}{Flag to sort the output. Default TR#'}
}
\value{
GRanges
}
\description{
Shortcut for \code{reduce(sort(gr.stripstrand(unlist(x))))}
}

