% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vaggregate}
\alias{vaggregate}
\title{vaggregate}
\usage{
vaggregate(...)
}
\arguments{
\item{...}{things to aggregate}
}
\description{
same as aggregate except returns named vector
with names as first column of output and values as second
}

