% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gUtils.R
\name{varcount}
\alias{varcount}
\title{varcount}
\usage{
varcount(bams, gr, min.mapq = 0, min.baseq = 20, max.depth = 500,
  indel = FALSE, ...)
}
\arguments{
\item{bams}{vector of bam paths to get var counts from}

\item{gr}{GenomicRanges specifying region to run}

\item{min.mapq}{Min mapping quality [Default 0]}

\item{min.baseq}{Min base-quality [Default 20]}

\item{max.depth}{Max coverage at base [Default 500]}

\item{indel}{[Default FALSE]}

\item{...}{= other args go to read.bam}
}
\description{
Wrapper around applyPileups
}
\details{
takes in vector of bam paths, GRanges corresponding to sites / territories to query,
and outputs a list with fields
$counts = 3D matrix of base counts
(A, C, G, T, N) x sites x bams subject to mapq and baseq thresholds
#'

(uses varbase)

... = other args go to read.bam
}

